% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{addPoints}
\alias{addPoints}
\title{Add points to a plotly object}
\usage{
addPoints(
  p,
  x,
  y,
  z = NULL,
  colorby = NULL,
  name = NULL,
  label = FALSE,
  plot_type = "scatter",
  point_size = 5,
  labels = NULL,
  showlegend = FALSE,
  palette = NULL
)
}
\arguments{
\item{p}{Previously generated plotly object}

\item{x}{Vector of numeric x values}

\item{y}{Vector of numeric y values}

\item{z}{Optional vector of numeric z values}

\item{colorby}{String vector or factor specifying value groups}

\item{name}{Name for the series}

\item{label}{Boolean- should points be colored and labelled?}

\item{plot_type}{Plot type: 'scatter' or 'scatter3d'}

\item{point_size}{Main point size}

\item{labels}{Vector of labels to apply (if 'label' is TRUE)}

\item{showlegend}{Boolean: show this set of points in the legend?}

\item{palette}{Colors used for grouped points}
}
\value{
output Plotly plot object
}
\description{
Add points to a plotly object
}
