% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{adjustLayout}
\alias{adjustLayout}
\title{Apply layout adjustments to plotly object}
\usage{
adjustLayout(
  p,
  title = "",
  legend_title = "",
  xlab = "x",
  ylab = "y",
  zlab = "z",
  plot_type = "scatter"
)
}
\arguments{
\item{p}{Previously generated plotly object}

\item{title}{Plot title}

\item{legend_title}{Legend title}

\item{xlab}{X axis label}

\item{ylab}{Y axis label}

\item{zlab}{Z axis label}

\item{plot_type}{Plot type: 'scatter' or 'scatter3d'}
}
\value{
output Plotly object
}
\description{
Apply layout adjustments to plotly object
}
