% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assaydatatable.R
\name{assaydatatable}
\alias{assaydatatable}
\title{The server function of the assaydatatable module}
\usage{
assaydatatable(id, eselist)
}
\arguments{
\item{id}{Module namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\description{
This function is called directly, using the same id as its UI counterpart,
and wraps its logic in \code{moduleServer()} (see example). Essentially this just passes the results of \code{colData()}
applied to the specified SummarizedExperiment object to the
\code{simpletable} module
}
\examples{
assaydatatable("assaydatatable", eselist)

# Almost certainly used via application creation

data(zhangneurons)
app <- prepareApp("assaydatatable", zhangneurons)
shiny::shinyApp(ui = app$ui, server = app$server)

}
\keyword{shiny}
