% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assaydatatable.R
\name{assaydatatableInput}
\alias{assaydatatableInput}
\title{The UI input function of the assaydatatable module}
\usage{
assaydatatableInput(id, eses)
}
\arguments{
\item{id}{Submodule namespace}

\item{eses}{List of ExploratorySummarizedExperiment objects with assay and
experimental data}
}
\value{
output An HTML tag object that can be rendered as HTML using 
  as.character()
}
\description{
This module displays the content of the currently selected experiment and 
assay, also allowing grouping by mean etc.
}
\details{
Leverages the \code{simpletable} module
}
\examples{
assaydatatableInput('experiment', eses)
}
\keyword{shiny}

