% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assaydatatable.R
\name{assaydatatableOutput}
\alias{assaydatatableOutput}
\title{The output function of the assaydatatable module}
\usage{
assaydatatableOutput(id)
}
\arguments{
\item{id}{Module namespace}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
This module produces a simple table of the \code{colData()} in a
SummarizedExperiment object. If more than one of these objects were 
specified, a select box should appear to pick the desired one for display.
}
\details{
Leverages the \code{simpletable} module
}
\examples{
assaydatatableOutput('experiment')
}
\keyword{shiny}

