% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesetbarcodeplot.R
\name{barcode_plot}
\alias{barcode_plot}
\title{Make a gene set barcode plot using Limma}
\usage{
barcode_plot(fold_changes, gene_ids, set_gene_ids, plot_title)
}
\arguments{
\item{fold_changes}{A list of fold changes}

\item{gene_ids}{Gene set IDs for the values in \code{fold_changes}}

\item{set_gene_ids}{Gene IDs for the gene set}

\item{plot_title}{A title for the plot}
}
\description{
Make a gene set barcode plot using Limma
}

