% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{bootstrap_median}
\alias{bootstrap_median}
\title{Bootstrap the standard error of the median}
\usage{
bootstrap_median(data, num)
}
\arguments{
\item{data}{Numeric vector}

\item{num}{Number of bootstraps}
}
\value{
out The bootstrap estimate of the median
}
\description{
Bootstrap the standard error of the median
}
\examples{
bootstrap_median(rnorm(50), num = 100)

}
