% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{bootstrapMedian}
\alias{bootstrapMedian}
\title{Bootstrap the standard error of the median}
\usage{
bootstrapMedian(data, num)
}
\arguments{
\item{data}{Numeric vector}

\item{num}{Number of bootstraps}
}
\value{
out The bootstrap estimate of the median
}
\description{
Bootstrap the standard error of the median
}
