% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplotOutput}
\alias{boxplotOutput}
\title{The output function of the boxplot module}
\usage{
boxplotOutput(id)
}
\arguments{
\item{id}{Submodule namespace}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
This provides actual boxplot element for display by applications
}
\examples{
boxplotOutput('boxplot')
}
\keyword{shiny}

