% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{build_enrichment_path}
\alias{build_enrichment_path}
\title{Build path to the enrichment results}
\usage{
build_enrichment_path(template, contrast_info, geneset_type, direction = NULL)
}
\arguments{
\item{template}{A string, such as \code{"/path/to/folder/{contrast_name}-{geneset_type}.csv"} or
\code{"./{contrast_name}/{geneset_type}/report_for_{target|reference}.csv"}}

\item{contrast_info}{A list with contrast details: `id`, `reference`, and `target`,
to be replaced in template.}

\item{geneset_type}{The name of the geneset type, to be replaced in the template}

\item{direction}{Either `"up"`, `"down"` or `NULL`, used to determine how the replacement will happen.}
}
\value{
A string similar to template, but with the templates replaced
}
\description{
Build path to the enrichment results
}
\details{
The template accepts the following:

\describe{
  \item{\code{\{contrast_name\}}}{Will be replaced by \code{contrast_info$id} argument}
  \item{\code{\{geneset_type\}}}{Will be replaced by the \code{geneset_type} argument}
  \item{\code{\{target|reference\}}}{If the \code{direction} argument is \code{"up"}, will be replaced
  with \code{contrast_info$target}, if it is \code{"down"}, \code{contrast_info$reference} will be used instead.}
}
}
\examples{
build_enrichment_path(
  template = "./{contrast_name}/{geneset_type}/report_for_{target|reference}.csv",
  contrast_info = list(id="disease_vs_ctrl", reference="control", target="disease"),
  geneset_type = "m2.cp.v2024.1.Mm.entrez",
  direction = "up"
)

}
