% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro.R
\name{calculateDendrogram}
\alias{calculateDendrogram}
\title{Calculate a clustering dendgrogram based on correlation}
\usage{
calculateDendrogram(
  plotmatrix,
  cor_method = "spearman",
  cluster_method = "ward.D2"
)
}
\arguments{
\item{plotmatrix}{Expression/ other data matrix}

\item{cor_method}{'spearman' or 'perason'}

\item{cluster_method}{Clustering method to pass to hclust (Default: 'ward.D2')}
}
\value{
output Object of class 'dist'
}
\description{
Calculate a clustering dendgrogram based on correlation
}
\examples{
data(airway, package = 'airway')
mymatrix <- assays(airway)[[1]]
calculateDendrogram(mymatrix)
}
\keyword{keywords}
