% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro.R
\name{calculateDist}
\alias{calculateDist}
\title{Calculate a distance matrix based on correlation}
\usage{
calculateDist(plotmatrix, cor_method = "spearman")
}
\arguments{
\item{plotmatrix}{Expression/ other data matrix}

\item{cor_method}{'spearman' or 'perason'}
}
\value{
output Object of class 'dist'
}
\description{
Calculate a distance matrix based on correlation
}
\examples{
data(airway, package = "airway")
mymatrix <- assays(airway)[[1]]
calculateDist(mymatrix)

}
\keyword{keywords}
