% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{calculatePCAFractionExplained}
\alias{calculatePCAFractionExplained}
\title{Extract the percent variance from a PCA analysis}
\usage{
calculatePCAFractionExplained(pca)
}
\arguments{
\item{pca}{An output from \code{prcomp}}
}
\value{
output vector of percentages
}
\description{
Extract the percent variance from a PCA analysis
}
\examples{
calculatePCAFractionExplained(pca)

}
\keyword{shiny}
