% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-string.R
\name{capitalize_first}
\alias{capitalize_first}
\title{Capitalise the first letter of a string}
\usage{
capitalize_first(string)
}
\arguments{
\item{string}{A string}
}
\value{
output String with capitalised first letter
}
\description{
Capitalise the first letter of a string
}
\examples{
capitalize_first("Example")

}
