% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorycountplot.R
\name{categorycountplotInput}
\alias{categorycountplotInput}
\title{The input function of the categorycountplot module}
\usage{
categorycountplotInput(id)
}
\arguments{
\item{id}{Submodule namespace}
}
\value{
A list of controls that can be added to a UI definition
}
\description{
Provides 'Count by'/'Split by' selectors and a bar mode toggle for tallying
rows of a data frame by a categorical column, plus an export table for the
resulting counts.
}
\examples{
categorycountplotInput("categorycount")

}
\keyword{shiny}
