% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorycountplot.R
\name{categorycountplotOutput}
\alias{categorycountplotOutput}
\title{The output function of the categorycountplot module}
\usage{
categorycountplotOutput(id)
}
\arguments{
\item{id}{Submodule namespace}
}
\value{
A list of elements that can be included in a panel
}
\description{
The output function of the categorycountplot module
}
\examples{
categorycountplotOutput("categorycount")

}
\keyword{shiny}
