% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts-validate.R
\name{check_list_is_subset}
\alias{check_list_is_subset}
\title{Check one list is a subset of another and throw an error if not}
\usage{
check_list_is_subset(
  test_list,
  reference_list,
  test_list_name,
  reference_list_name
)
}
\arguments{
\item{test_list}{Test list}

\item{reference_list}{Reference list}

\item{test_list_name}{Name of test list for error}

\item{reference_list_name}{Name of reference list for error}
}
\value{
output Returns TRUE if check passes
}
\description{
Check one list is a subset of another and throw an error if not
}
\examples{
check_list_is_subset(c("treatment"), c("treatment", "batch"), "contrast variables", "sample metadata")

}
