% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipseq.R
\name{chipseq}
\alias{chipseq}
\title{The server function of the chipseq module}
\usage{
chipseq(id, eselist, heatmap_layout = heatmap_layout_options())
}
\arguments{
\item{id}{Module namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}

\item{heatmap_layout}{A list of heatmap pixel-size options as produced by
\code{\link{heatmap_layout_options}}, passed through to every heatmap
panel in the app}
}
\description{
This function is called directly, using the same id as its UI counterpart,
and wraps its logic in \code{moduleServer()} (see example).
}
\examples{
chipseq("chipseq", eselist)

}
\keyword{shiny}
