% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipseq.R
\name{chipseqInput}
\alias{chipseqInput}
\title{The input function of the chipseq module}
\usage{
chipseqInput(id, eselist)
}
\arguments{
\item{id}{Submodule namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\value{
output An HTML tag object that can be rendered as HTML using
  as.character()
}
\description{
This provides the form elements to control the ChIP-seq display. The chipseq
module combines output from analysis modules such as PCA and boxplots into a
comprehensive application.
}
\examples{
chipseqInput("chipseq", eselist)

}
\keyword{shiny}
