% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{chooseGroupingVariables}
\alias{chooseGroupingVariables}
\title{Choose a valid set of grouping variables from a targets/ experiment data
frame.}
\usage{
chooseGroupingVariables(df)
}
\arguments{
\item{df}{Input data frame}
}
\value{
out A list of valid column names
}
\description{
To be useful for grouping a variable needs to be a character or a factor, 
and to have a number of values greater than one but less than the number of
samples.
}
\examples{
# `airway` contains info on the samples it's based on

data(airway, package = 'airway')

# However, not all variables are useful for grouping data. Some have a 
# different value for every sample, one has the same value for all of them.

colData(airway)

# So just pick the variables that ARE useful

chooseGroupingVariables(data.frame(colData(airway)))
}
