% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering}
\alias{clustering}
\title{The server function of the clustering module}
\usage{
clustering(id, eselist)
}
\arguments{
\item{id}{Module namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\description{
This module plots the expression profiles (scaled for comparison) of the
selected rows of the input matrix provided by the \code{selectmatrix} module.
}
\details{
The \code{\link[clustering]{clara}} method, a fast approximation of
partitioning about medoids, is used to produce the clusters. As well as
defining the input matrix users can decide how the clusters are drawn and how
many clusters should be generated.

This function is called directly, using the same id as its UI counterpart,
and wraps its logic in \code{moduleServer()} (see example).
}
\examples{
clustering("myid", eselist)

}
\keyword{shiny}
