% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro.R
\name{clustering_dendrogram}
\alias{clustering_dendrogram}
\title{Make a clustering dendrogram with coloring by experimental variable}
\usage{
clustering_dendrogram(plotmatrix, experiment, colorby = NULL,
  corMethod = "pearson", clusterMethod = "ward.D", plot_title = "")
}
\arguments{
\item{plotmatrix}{Expression/ other data matrix}

\item{experiment}{Annotation for the columns of plotmatrix}

\item{colorby}{Column name in \code{experiment} specifying how boxes should be colored}
}
\value{
output A \code{ggplot} output
}
\description{
A simple function using \code{ggdendro} to make a sample dendrogram
}
\examples{
ggplot_boxplot(selectMatrix(), selectColData(), colorBy())
}
\keyword{keywords}

