% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clusteringOutput}
\alias{clusteringOutput}
\title{The output function of the clustering module}
\usage{
clusteringOutput(id)
}
\arguments{
\item{id}{Submodule namespace}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
This module plots the expression profiles (scaled for comparison) of the 
selected rows of the input matrix provided by the \code{selectmatrix} module.
}
\details{
The \code{\link[clustering]{clara}} method, a fast approximation of 
partitioning about medoids, is used to produce the clusters. As well as 
defining the input matrix users can decide how the clusters are drawn and how
many clusters should be generated.

This funcion provides the form elements to control the display

This provides actual plot element for display by applications
}
\examples{
clusteringOutput('myid')
}
\keyword{shiny}
