% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarisematrix.R
\name{col_geom_means}
\alias{col_geom_means}
\title{Geometric means by matrix column}
\usage{
col_geom_means(x)
}
\arguments{
\item{x}{A matrix}
}
\value{
Vector with column geometric means
}
\description{
Geometric means by matrix column
}
\examples{
m <- matrix(1:6, nrow = 2, dimnames = list(NULL, c("a", "b", "c")))
col_geom_means(m)

}
