% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarisematrix.R
\name{col_medians}
\alias{col_medians}
\title{Medians by matrix column}
\usage{
col_medians(x)
}
\arguments{
\item{x}{A matrix}
}
\value{
Vector with column medians
}
\description{
Medians by matrix column
}
\examples{
m <- matrix(1:6, nrow = 2, dimnames = list(NULL, c("a", "b", "c")))
col_medians(m)

}
