% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarisematrix.R
\name{colMedians}
\alias{colMedians}
\title{Medians by matrix column}
\usage{
colMedians(x)
}
\arguments{
\item{x}{A matrix}
}
\value{
Vector with column medians
}
\description{
Medians by matrix column
}
