% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colormaker.R
\name{colormakerInput}
\alias{colormakerInput}
\title{The input function of the colorby module}
\usage{
colormakerInput(id)
}
\arguments{
\item{id}{Submodule namespace}
}
\value{
output An HTML tag object that can be rendered as HTML using
  as.character()
}
\description{
Provides a drop-down for picking the plot colour palette. The default is
shinyngs' fixed colour-blind-safe palette (see \code{\link{make_color_scale}});
the remaining choices are RColorBrewer qualitative palettes for users who
prefer them.
}
\details{
This function provides the form elements to control the display
}
\examples{
colormakerInput("myid")

}
\keyword{shiny}
