% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{compile_pca_data}
\alias{compile_pca_data}
\title{Run PCA on a given matrix, expected to be variance stabilised (at least
log-transformed)}
\usage{
compile_pca_data(matrix, ntop = NULL, scale_features = FALSE)
}
\arguments{
\item{matrix}{Simple matrix with genes by row and samples by column}

\item{ntop}{Number of most variable genes to use}

\item{scale_features}{Boolean, passed through to \code{\link{runPCA}}'s
\code{scale_features} argument. Defaults to \code{FALSE}, appropriate for
the variance-stabilised matrices this function is normally called with.}
}
\value{
a list with keys 'coords' and 'percentVar' providing PCA coordinates
  and fractional variance contributions, respectively.
}
\description{
Run PCA on a given matrix, expected to be variance stabilised (at least
log-transformed)
}
\examples{
mat <- matrix(rnorm(60), nrow = 15, ncol = 4,
  dimnames = list(paste0("gene", 1:15), paste0("s", 1:4)))
pca <- compile_pca_data(mat)
head(pca$coords)

}
