% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{compilePCAData}
\alias{compilePCAData}
\title{Run PCA on a given matrix, expected to be variance stabilised (at least
log-transformed)}
\usage{
compilePCAData(matrix, ntop = opt$n_genes)
}
\arguments{
\item{matrix}{Simple matrix with genes by row and samples by column}

\item{ntop}{Number of most variable genes to use}
}
\value{
a list with keys 'coords' and 'percentVar' providing PCA coordinates
  and fractional variance contributions, respectively.
}
\description{
Run PCA on a given matrix, expected to be variance stabilised (at least
log-transformed)
}
