% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{contrasts}
\alias{contrasts}
\title{The server function of the contrasts module}
\usage{
contrasts(
  id,
  eselist,
  selectmatrix_reactives = list(),
  multiple = FALSE,
  select_all_contrasts = FALSE,
  show_controls = TRUE,
  summarise = TRUE,
  default_foldchange = 2,
  default_pval = 0.05,
  default_qval = 0.1
)
}
\arguments{
\item{id}{Module namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}

\item{selectmatrix_reactives}{The list of reactive expressions returned by
a call to the \code{\link{selectmatrix}} module. This will be unpacked to
gain access to the data provided by those reaactive.}

\item{multiple}{Allow selection of multiple contrasts?}

\item{select_all_contrasts}{Select all contrasts by default?}

\item{show_controls}{Show the controls for contrast selection?}

\item{summarise}{Match the corresponding \code{contrastsInput()} setting.
When `TRUE`, wait for the summary-type input before producing contrast data.}

\item{default_foldchange}{default value for the fold change filter}

\item{default_pval}{Default value for the p value field}

\item{default_qval}{Default value for the q value field}
}
\description{
This module provides the form elements to control contrasts used in e.g.
differential expression panels. In particular it provides the ability for
users to add filters to progressively refine a query.
}
\details{
This function is called directly, using the same id as its UI counterpart,
and wraps its logic in \code{moduleServer()} (see example).

The reactive graph is built by a sequence of internal factory functions,
each owning one cohesive stage (contrast enumeration, naming, the dynamic
filter-set engine, table building, selection, filtering, labelling, and
the query summary) - see \code{contrastEnumeration}, \code{contrastNaming},
\code{contrastFilterSetEngine}, \code{contrastTableBuilder},
\code{contrastSelection}, \code{contrastFiltering}, \code{contrastLabelling}
and \code{contrastQuerySummary}. This function itself just wires those
stages together in dependency order and returns their public reactives.
}
\examples{
contrasts("differential", eselist = eselist, selectmatrix_reactives = selectmatrix_reactives, multiple = TRUE)

}
\keyword{shiny}
