% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{contrastsOutput}
\alias{contrastsOutput}
\title{The output function of the contrasts module}
\usage{
contrastsOutput(id)
}
\arguments{
\item{id}{Submodule namespace}
}
\description{
This module provides the form elements to control contrasts used in e.g.
differential expression panels. In particular it provides the ability for
users to add filters to progressively refine a query.
}
\details{
This function provides a summary. Actual output should be rendered
by calling modules using the provided reactives.
}
\examples{
contrastsOutput("myid")

}
