% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-string.R
\name{count_lines}
\alias{count_lines}
\title{Count the number of lines in a string}
\usage{
count_lines(string)
}
\arguments{
\item{string}{A string}
}
\value{
output Integer supplying the number of lines in a string
}
\description{
Count the number of lines in a string
}
\examples{
count_lines("foo\nbar")
}
