% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro.R
\name{dendro}
\alias{dendro}
\title{The server function of the dendrogram module}
\usage{
dendro(id, eselist)
}
\arguments{
\item{id}{Module namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\description{
This module will produce a sample clustering dendrogram based on
user-selected parameters of row (e.g. gene) and column (sample) selection
provided by the \code{selectmatrix} module, as well distance matrix
generation and clustering method.
}
\details{
This function is called directly, using the same id as its UI counterpart,
and wraps its logic in \code{moduleServer()} (see example).
}
\examples{
dendro("myid", eselist)

}
\keyword{shiny}
