% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro.R
\name{dendroInput}
\alias{dendroInput}
\title{The input function of the dendrogram module}
\usage{
dendroInput(id, ses)
}
\arguments{
\item{id}{Submodule namespace}

\item{ses}{List of structuredExperiment objects with assay and experimental
data, with additional information in the metadata() slot}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
This provides the form elements to control the pca display
}
\examples{
dendroInput(ns('boxplot'), ses)
}
\keyword{shiny}

