% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro.R
\name{dendroOutput}
\alias{dendroOutput}
\title{The output function of the boxplot module}
\usage{
dendroOutput(id)
}
\arguments{
\item{id}{Submodule namespace}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
This provides actual boxplot element for display by applications
}
\examples{
dendroOutput('dendro')
}
\keyword{shiny}

