% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexseqtable.R
\name{dexseqtableInputFields}
\alias{dexseqtableInputFields}
\title{Make input fields for producing a table of differential exon usage.
Separated here for re-use by the dexseqplot module}
\usage{
dexseqtableInputFields(id, eselist, allow_filtering = TRUE)
}
\arguments{
\item{id}{Submodule namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}

\item{allow_filtering}{Allow user filtering of results (default: TRUE)?
Deactivated by the \code{dexseqplot} module, which uses it for showing
gene-specific results.}
}
\value{
output Named list of \code{shiny.tag} objects
}
\description{
Make input fields for producing a table of differential exon usage.
Separated here for re-use by the dexseqplot module
}

