% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialsummary.R
\name{differentialsummary}
\alias{differentialsummary}
\title{Server function for the differential summary module}
\usage{
differentialsummary(id, eselist)
}
\arguments{
\item{id}{Module namespace.}

\item{eselist}{ExploratorySummarizedExperimentList object.}
}
\description{
Server function for the differential summary module
}
\keyword{shiny}
