% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialsummary.R
\name{differentialsummaryInput}
\alias{differentialsummaryInput}
\title{Input function for the differential summary module}
\usage{
differentialsummaryInput(id, eselist)
}
\arguments{
\item{id}{Module namespace.}

\item{eselist}{ExploratorySummarizedExperimentList object.}
}
\value{
Controls for experiment, assay, differential filters, and export.
}
\description{
Input function for the differential summary module
}
\keyword{shiny}
