% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialtable.R
\name{differentialtable}
\alias{differentialtable}
\title{The server function of the differentialtable module}
\usage{
differentialtable(id, eselist)
}
\arguments{
\item{id}{Module namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\description{
This module provides information on the comparison betwen pairs of groups
defined in a 'contrasts' slot of a ExploratorySummarizedExperimentList
}
\details{
This function is called directly, using the same id as its UI counterpart,
and wraps its logic in \code{moduleServer()} (see example).

Essentially this funnction uses the \code{contrasts} module to group samples
and calculate fold changes, adding test statistics where available.
}
\examples{
differentialtable("differentialtable", eselist)

}
\keyword{shiny}
