% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialtable.R
\name{differentialtableOutput}
\alias{differentialtableOutput}
\title{The output function of the differentialtable module}
\usage{
differentialtableOutput(id)
}
\arguments{
\item{id}{Module namespace}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
This module provides information on the comparison betwen pairs of groups 
defined in a 'contrasts' slot of a ExploratorySummarizedExperimentList
}
\details{
Leverages the \code{simpletable} module
}
\examples{
differentialtableOutput('experiment')
}
\keyword{shiny}

