% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{drawLines}
\alias{drawLines}
\title{Overlay lines on a plotly-generated plot}
\usage{
drawLines(
  p,
  x,
  y,
  lines = NULL,
  hline_thresholds = list(),
  vline_thresholds = list()
)
}
\arguments{
\item{p}{Previously generated plotly object}

\item{x}{X coordinates of points, used to determine x range}

\item{y}{Y coordinates of points, used to determine y range}

\item{lines}{3 column data-frame (name, x, y) with two rows, one for the
start and end of each named line}

\item{hline_thresholds}{Alternatively or in addition to 'lines', just specify
a named list of y values at which to place hlines}

\item{vline_thresholds}{Alternatively or in addition to 'lines', just specify
a named list of x values at which to place vlines}
}
\value{
output Plotly object
}
\description{
Overlay lines on a plotly-generated plot
}
