% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentoverview.R
\name{enrichmentoverview}
\alias{enrichmentoverview}
\title{Server function for the gene set overview module}
\usage{
enrichmentoverview(id, eselist)
}
\arguments{
\item{id}{Module namespace.}

\item{eselist}{ExploratorySummarizedExperimentList object.}
}
\description{
Server function for the gene set overview module
}
\keyword{shiny}
