% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentoverview.R
\name{enrichmentoverviewInput}
\alias{enrichmentoverviewInput}
\title{Input function for the gene set overview module}
\usage{
enrichmentoverviewInput(id, eselist)
}
\arguments{
\item{id}{Module namespace.}

\item{eselist}{ExploratorySummarizedExperimentList object.}
}
\value{
Controls for the experiment, assay, gene set type, ranking, and export.
}
\description{
Input function for the gene set overview module
}
\keyword{shiny}
