% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentoverview.R
\name{enrichmentoverviewOutput}
\alias{enrichmentoverviewOutput}
\title{Output function for the gene set overview module}
\usage{
enrichmentoverviewOutput(id)
}
\arguments{
\item{id}{Module namespace.}
}
\value{
Outputs for the method label, plot, and backing table.
}
\description{
Output function for the gene set overview module
}
\keyword{shiny}
