% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{ensureLog}
\alias{ensureLog}
\title{Ensure a matrix is in log scale}
\usage{
ensureLog(vals, condition, rmzeros = FALSE)
}
\arguments{
\item{vals}{Numberic matrix or vector}

\item{condition}{Boolean- apply log transform?}

\item{rmzeros}{Remove zeros before applying transform?}
}
\value{
output Object of same type as vals with tranform applied where
  specified
}
\description{
Ensure a matrix is in log scale
}
