% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{eselistFromYAML}
\alias{eselistFromYAML}
\title{Build an ExploratorySummarisedExperimentList from a YAML description}
\usage{
eselistFromYAML(configfile)
}
\arguments{
\item{configfile}{A YAML-format config file describing the data to be
compiled into an ExploratorySummarizedExperimentList object}
}
\value{
out An ExploratorySummarizedExperimentList object suitable for passing to \code{\link{prepareApp}}
}
\description{
Building ExploratorySummarisedExperimentList objects can be a bit fiddly.
This function makes automates object construction based on a descriptor
in yaml format.
}
\details{
For a simple study with one 'experiement' for Gene-level results, and three
'assays' describing raw, filtered and normalised expression you might make
a YAML like:

\if{html}{\out{<div class="sourceCode">}}\preformatted{title: My RNA seq experiment
author: Joe Blogs
report: report.md
group_vars:
  - Group
  - Replicate
default_groupvar: Group
experiments:
  Gene:
    coldata:
      file: my.experiment.csv
      id: External
    annotation:
      file: my.annotation.csv
      id: gene_id
      entrez: ~
      label: gene_id
    expression_matrices:
      Raw:
        file: raw_counts.csv
        measure: counts
      Filtered:
        file: filtered_counts.csv
        measure: Counts per million
      Normalised:
        file: normalised_counts.csv
        measure: Counts per million
    read_reports:
      read_attrition: read_attrition.csv
contrasts:
  comparisons:
  - Variable: Group
    Group.1: control
    Group.2: TreatmentA
  - Variable: Group
    Group.1: control
    Group.2: TreatmentB
contrast_stats:
  Gene:
    Normalised:
      pvals: pvals.csv
      qvals: qvals.csv
}\if{html}{\out{</div>}}
}
\examples{
eselist <- eselistFromYAML("my.yaml")

}
