% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{evaluateCardinalFilter}
\alias{evaluateCardinalFilter}
\title{Evaluate a vector of values with respect to a limit and a cardinality, being
'>', '<' , '> or <-' (e.g. a fold change above a limit in + or - 
directions), or '< and >-' (not a above a limit in + or -).}
\usage{
evaluateCardinalFilter(values, cardinality, limit)
}
\arguments{
\item{values}{}

\item{cardinality}{}

\item{limit}{}
}
\value{
out A logical vector
}
\description{
Evaluate a vector of values with respect to a limit and a cardinality, being
'>', '<' , '> or <-' (e.g. a fold change above a limit in + or - 
directions), or '< and >-' (not a above a limit in + or -).
}
