% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-from-mae.R
\name{ExploratorySummarizedExperiment.SingleFact}
\alias{ExploratorySummarizedExperiment.SingleFact}
\title{Creates an object of the ExploratorySummarizedExperiment class}
\usage{
ExploratorySummarizedExperiment.SingleFact(
  assays,
  colData,
  annotation,
  idfield,
  labelfield = character(),
  entrezgenefield = character(),
  contrast_stats = list(),
  assay_measures = list(),
  gene_set_analyses = list(),
  dexseq_results = list(),
  read_reports = list()
)
}
\arguments{
\item{See}{?shinyngs::ExploratorySummarizedExperiment}
}
\description{
This is fixes a bug in \code{shinyngs::ExploratorySummarizedExperiment}; the container \code{colData} stores the phenotypic data and if there is only a single factor (i.e. one column), it gets converted to a vector & IS NOT preserved as a dataframe. Hence assignment of 'assays' fails and throws an error because the sample names have been lost! Setting "drop=FALSE" when manipulating this object prevents this behaviour.
}
