% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-string.R
\name{file_extension}
\alias{file_extension}
\title{Extract the extension of a file}
\usage{
file_extension(file)
}
\arguments{
\item{file}{Input file path}
}
\value{
extension Extension like 'csv' or 'tsv'
}
\description{
Extract the extension of a file
}
\examples{
file_extension("my_metadata.tsv")

}
