% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts-validate.R
\name{fixedEffectsFormula}
\alias{fixedEffectsFormula}
\title{Remove random effects from a model formula}
\usage{
fixedEffectsFormula(formula_string)
}
\arguments{
\item{formula_string}{Formula string that may contain random effects}
}
\value{
output Fixed-effects-only formula
}
\description{
Remove random effects from a model formula
}
