% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{fixedEffectsModelMatrix}
\alias{fixedEffectsModelMatrix}
\title{Build a model matrix from the fixed-effects part of a formula}
\usage{
fixedEffectsModelMatrix(formula_string, samples)
}
\arguments{
\item{formula_string}{Formula string that may contain random effects}

\item{samples}{Data frame of sample information}
}
\value{
output Model matrix for the fixed-effects formula
}
\description{
Build a model matrix from the fixed-effects part of a formula
}
