% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{fold_change}
\alias{fold_change}
\title{Calculate fold change between two vectors}
\usage{
fold_change(vec1, vec2)
}
\arguments{
\item{vec1}{First vector}

\item{vec2}{Second vector}
}
\value{
Vector of fold changes
}
\description{
Calculate fold change between two vectors
}
\examples{
fold_change(c(2, 2, 2), c(2, 4, 8))

}
