% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{foldChange}
\alias{foldChange}
\title{Calculate fold change between two vectors}
\usage{
foldChange(vec1, vec2)
}
\arguments{
\item{vec1}{First vector}

\item{vec2}{Second vector}
}
\value{
Vector of fold changes
}
\description{
Calculate fold change between two vectors
}
\examples{
foldChange(c(2,2,2), c(2,4,8))
}

