% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldchangeplot.R
\name{foldchangeplot}
\alias{foldchangeplot}
\title{The server function of the \code{foldchangeplot} module}
\usage{
foldchangeplot(id, eselist)
}
\arguments{
\item{id}{Module namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\description{
This module is for making scatter plots comparing pairs of groups defined in
a 'contrasts' slot of the ExploratorySummarizedExperimentList
}
\details{
This function is called directly, using the same id as its UI counterpart,
and wraps its logic in \code{moduleServer()} (see example).
}
\examples{
data(airway, package = "airway")
ese <- as(airway, "ExploratorySummarizedExperiment")
eselist <- ExploratorySummarizedExperimentList(ese)

if (interactive()) {
  foldchangeplot("foldchangeplot", eselist)
  app <- prepareApp("foldchangeplot", eselist)
  shiny::shinyApp(ui = app$ui, server = app$server)
}

}
\keyword{shiny}
