% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{geneBarplot}
\alias{geneBarplot}
\title{Main function for drawing the bar plot with plotly}
\usage{
geneBarplot(expression, experiment, colorby, expressionmeasure = "Expression",
  palette = NULL)
}
\arguments{
\item{expression}{Matrix of values}

\item{experiment}{Data frame containing metadata to use for coloring etc}

\item{colorby}{Column name in \code{experiment} specifying how points should be colored}

\item{expressionmeasure}{String to use for labelling y axis in plots}

\item{palette}{Palette of colors, one for each unique value derived from 
\code{colorby}.}
}
\description{
The gene module picks specified rows out the assay data, either simply by 
id or label. This is used to create a gene-centric info page.
}
\details{
This function does the actual plotting with plotly. It will produce a plot 
for every row of the input matrix.
}
\examples{
callModule(gene, 'gene', ses)
}
\keyword{shiny}
