% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{geneInput}
\alias{geneInput}
\title{The input function of the gene module}
\usage{
geneInput(id, eselist)
}
\arguments{
\item{id}{Submodule namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\value{
output An HTML tag object that can be rendered as HTML using 
  as.character()
}
\description{
The gene module picks specified rows out the assay data, either simply by id
or label. This is used to create a gene-centric info page.
}
\details{
Inputs are a gene label and a variable to color by (where available)
}
\examples{
geneInput(ns('gene'), eselist)
}
\keyword{shiny}
