% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{geneModelPlot}
\alias{geneModelPlot}
\title{Make a gene model plot for a chromosomal location}
\usage{
geneModelPlot(ensembl_species, chromosome, start, end)
}
\arguments{
\item{ensembl_species}{}

\item{chromosome}{}

\item{start}{}

\item{end}{}
}
\description{
Make a gene model plot for a chromosomal location
}

