% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\name{geneset}
\alias{geneset}
\title{The server function of the gene set module}
\usage{
geneset(input, output, session, eselist, getExperiment)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{eselist}{An ExploratorySummarizedExperimentList with its gene_sets
slot set}

\item{getExperiment}{Accessor for returning an
ExploratorySummarizedExperiment object, with 'entrezgenefield', 
'labelfield' set in its slots}
}
\value{
output A list of two reactive functions: getPathwayNames() and 
  getPathwayGenes() which will be used by other modules.
}
\description{
The gene set module is for adding a gene set filter to shiny UI elements. 
This function is not called directly, but rather via callModule() (see 
example).
}
\details{
This function assumes that the gene sets have one gene ID (e.g. Entrez) which
need to be converted to another (e.g. Symbol, Ensembl gene ID). This would be
common when dealign with MSigDB gene sets, for example.
}
\examples{
geneset_functions <- callModule(geneset, 'heatmap', getExperiment())
}
\keyword{shiny}

