% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesetanalysistable.R
\name{genesetanalysistable}
\alias{genesetanalysistable}
\title{The server function of the genesetanalysistable module}
\usage{
genesetanalysistable(id, eselist)
}
\arguments{
\item{id}{Module namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\description{
This module displays gene set analysis tables stored as a list in the
\code{gene_set_analyses} slot of an \code{ExploratorySummarizedExperiment}.
}
\details{
The \code{gene_set_analyses} slot must be keyed first by the name of the
assay to which it pertains, and second by the gene set type (e.g. 'KEGG').
The containing \code{ExploratorySummarizedExperiment} must have a populated
\code{gene_sets} slot, keyed first by metadata column used to define the
gene sets and secondly by the gene set type.

The module is based on the output of roast() from \code{limma}, but it's
fairly generic, and assumes only the presence of a 'p value' and 'FDR'
column, so the output of other methods should be easily adapted to suit.

This function is called directly, using the same id as its UI counterpart,
and wraps its logic in \code{moduleServer()} (see example). Essentially this just passes the results of \code{colData()}
applied to the specified SummarizedExperiment object to the
\code{simpletable} module
}
\examples{

data(airway, package = "airway")
ese <- as(airway, "ExploratorySummarizedExperiment")
eselist <- ExploratorySummarizedExperimentList(ese)

# This module needs an eselist whose experiments carry gene_set_analyses
# results (see the vignette). Given those, the module server is called like:

if (interactive()) {
  genesetanalysistable("genesetanalysistable", eselist)
}

}
\keyword{shiny}
