% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesetanalysistable.R
\name{genesetanalysistableInput}
\alias{genesetanalysistableInput}
\title{The UI input function of the genesetanalysistable module}
\usage{
genesetanalysistableInput(id, eselist)
}
\arguments{
\item{id}{Submodule namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\value{
output An HTML tag object that can be rendered as HTML using
  as.character()
}
\description{
This module displays gene set analysis tables stored as a list in the
\code{gene_set_analyses} slot of an \code{ExploratorySummarizedExperiment}.
}
\details{
The \code{gene_set_analyses} slot must be keyed first by the name of the
assay to which it pertains, and second by the gene set type (e.g. 'KEGG').
The containing \code{ExploratorySummarizedExperiment} must have a populated
\code{gene_sets} slot, keyed first by metadata column used to define the
gene sets and secondly by the gene set type.

The module is based on the output of roast() from \code{limma}, but it's
fairly generic, and assumes only the presence of a 'p value' and 'FDR'
column, so the output of other methods should be easily adapted to suit.

Leverages the \code{simpletable} module
}
\examples{
data(airway, package = "airway")
ese <- as(airway, "ExploratorySummarizedExperiment")
eselist <- ExploratorySummarizedExperimentList(ese)

# This module needs an eselist whose experiments carry gene_set_analyses
# results (see the vignette). Given those, module input is produced like:

if (interactive()) {
  genesetanalysistableInput("experiment", eselist)
}

}
\keyword{shiny}
