% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesetbarcodeplot.R
\name{genesetbarcodeplot}
\alias{genesetbarcodeplot}
\title{The server function of the genesetbarcodeplot module}
\usage{
genesetbarcodeplot(id, eselist)
}
\arguments{
\item{id}{Module namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\description{
This module leverages gene sets stored in the \code{gene_sets} slot of an
\code{ExploratorySummarizedExperimentList} object to produce interactive
barcode plots, reproducing the statistics behind Limma's
\code{\link[limma]{barcodeplot}} function. Genes are ranked by fold changes
calculated with the \code{contrasts} module, and FDR values from the
\code{gene_set_analyses} slot of the selected
\code{ExploratorySummarizedExperiment} are displayed where provided.
}
\details{
This function is called directly, using the same id as its UI counterpart,
and wraps its logic in \code{moduleServer()} (see example).
}
\examples{
if (interactive()) {
  data(airway, package = "airway")
  ese <- as(airway, "ExploratorySummarizedExperiment")
  eselist <- ExploratorySummarizedExperimentList(ese)
  genesetbarcodeplot("genesetbarcodeplot", eselist)

  # Almost certainly used via application creation

  app <- prepareApp("genesetbarcodeplot", eselist)
  shiny::shinyApp(ui = app$ui, server = app$server)
}

}
\keyword{shiny}
