% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesetbarcodeplot.R
\name{genesetbarcodeplotInput}
\alias{genesetbarcodeplotInput}
\title{The UI input function of the genesetbarcodeplot module}
\usage{
genesetbarcodeplotInput(id, eselist)
}
\arguments{
\item{id}{Submodule namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\value{
output An HTML tag object that can be rendered as HTML using
  as.character()
}
\description{
This module leverages gene sets stored in the \code{gene_sets} slot of an
\code{ExploratorySummarizedExperimentList} object to produce interactive
barcode plots, reproducing the statistics behind Limma's
\code{\link[limma]{barcodeplot}} function. Genes are ranked by fold changes
calculated with the \code{contrasts} module, and FDR values from the
\code{gene_set_analyses} slot of the selected
\code{ExploratorySummarizedExperiment} are displayed where provided.
}
\examples{
data(airway, package = "airway")
ese <- as(airway, "ExploratorySummarizedExperiment")
eselist <- ExploratorySummarizedExperimentList(ese)

# The gene set modules need an eselist with populated gene_set_analyses (see
# the vignette). Given such data, they are used via application creation:

if (interactive()) {
  genesetbarcodeplotInput("myid", eselist)
  app <- prepare_app("genesetbarcodeplot", eselist)
  shiny::shinyApp(ui = app$ui, server = app$server)
}

}
\keyword{shiny}
