% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesetbarcodeplot.R
\name{genesetbarcodeplotOutput}
\alias{genesetbarcodeplotOutput}
\title{The output function of the genesetbarcodeplot module}
\usage{
genesetbarcodeplotOutput(id)
}
\arguments{
\item{id}{Module namespace}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
This module provides information on the comparison betwen pairs of groups 
defined in a 'tests' slot of the ExploratorySummarizedExperiment
}
\details{
Leverages the \code{simpletable} module
}
\examples{
genesetbarcodeplotOutput('experiment')
}
\keyword{shiny}

