% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\name{genesetInput}
\alias{genesetInput}
\title{The UI function of the gene set module}
\usage{
genesetInput(id)
}
\arguments{
\item{id}{Submodule namespace}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
The gene set module is for adding a gene set filter to shiny UI elements.
This will generally not be called directly, but by other modules
}
\examples{
genesetInput(ns("heatmap"))
}
\keyword{shiny}

