% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesetselect.R
\name{genesetselectInput}
\alias{genesetselectInput}
\title{The UI function of the genesetselect module}
\usage{
genesetselectInput(id, multiple = TRUE)
}
\arguments{
\item{id}{Submodule namespace}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
The gene set module is for adding a gene set filter to shiny UI elements.
This will generally not be called directly, but by other modules
}
\examples{
genesetselectInput(ns('heatmap'))
}
\keyword{shiny}

