% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarisematrix.R
\name{geom_mean}
\alias{geom_mean}
\title{Geometric mean}
\usage{
geom_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{A vector}

\item{na.rm}{Passed to \code{sum()}}
}
\value{
A numeric value
}
\description{
Geometric mean
}
