% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{getExtension}
\alias{getExtension}
\title{Extract the extension of a file}
\usage{
getExtension(file)
}
\arguments{
\item{file}{Input file path}
}
\value{
extension Extension like 'csv' or 'tsv'
}
\description{
Extract the extension of a file
}
