% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{getSeparator}
\alias{getSeparator}
\title{Infer a separator from the extension of an input file}
\usage{
getSeparator(file)
}
\arguments{
\item{file}{Input file path}
}
\value{
output Separator character like tab or ','
}
\description{
Infer a separator from the extension of an input file
}
