% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupby.R
\name{groupby}
\alias{groupby}
\title{The server function of the groupby module}
\usage{
groupby(
  id,
  eselist,
  group_label = "Group by",
  multiple = FALSE,
  selectColData = NULL,
  isDynamic = reactive({
     TRUE
 }),
  color = TRUE
)
}
\arguments{
\item{id}{Module namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}

\item{group_label}{A label for the grouping field}

\item{multiple}{Produces a checkbox group if true, a select box if false}

\item{selectColData}{Reactive returning an experiment matrix, probably
derived from the \code{\link{selectmatrix}} module.}

\item{isDynamic}{Reactive expression providing a boolean. A FALSE value
causes the groupby option to be placed in a hidden field.}

\item{color}{Require and return palette controls?}
}
\value{
output A list of reactive functions which will be used by other
modules.
}
\description{
The groupby module provides a UI element to choose from the
\code{group_vars} in a SummarizedExperment. Useful for coloring in a PCA etc
}
\examples{
geneset_functions <- groupby("heatmap", getExperiment)

}
\keyword{shiny}
