% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-string.R
\name{guess_separator}
\alias{guess_separator}
\title{Infer a separator from the extension of an input file}
\usage{
guess_separator(file)
}
\arguments{
\item{file}{Input file path}
}
\value{
output Separator character like tab or ','
}
\description{
Infer a separator from the extension of an input file
}
\examples{
guess_separator("my_metadata.tsv")

}
