% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{has_slot_data}
\alias{has_slot_data}
\title{Check whether a list-type slot on an S4 object is populated}
\usage{
has_slot_data(x, slot_name)
}
\arguments{
\item{x}{An S4 object (e.g. \code{ExploratorySummarizedExperiment} or
\code{ExploratorySummarizedExperimentList})}

\item{slot_name}{Name of the slot to check}
}
\value{
output Boolean- is the slot populated?
}
\description{
Centralises the \code{length(x@some_slot) > 0} idiom used in several Shiny
modules to decide whether optional data (contrasts, gene sets, read
reports, DEXSeq results etc) is available before showing or hiding UI for
it.
}
\examples{
data(airway, package = "airway")
ese <- as(airway, "ExploratorySummarizedExperiment")
eselist <- ExploratorySummarizedExperimentList(ese)
has_slot_data(eselist, "contrasts")

}
\keyword{shiny}
