% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap}
\alias{heatmap}
\title{The server function of the heatmap module}
\usage{
heatmap(
  id,
  eselist,
  type = "expression",
  heatmap_layout = heatmap_layout_options()
)
}
\arguments{
\item{id}{Module namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}

\item{type}{The type of heatmap that will be made. 'expression', 'samples' or
'pca' (default: 'expression')}

\item{heatmap_layout}{A list of heatmap pixel-size options as produced by
\code{\link{heatmap_layout_options}}}
}
\description{
Three types of heatmaps are provided, employed in various places in the
rnaseq app (for example), and using much of the same code. Expresssion
heatmaps plot expression for samples by column and e.g. genes by row. A
samples heatmap plots samples vs samples to illustrate correlation patterns.
A pca heatmap plots the results of anova tests applied to examine the
associations between principal components and experimental variables.
}
\details{
This function is called directly, using the same id as its UI counterpart,
and wraps its logic in \code{moduleServer()} (see example).
}
\examples{
data(airway, package = "airway")
ese <- as(airway, "ExploratorySummarizedExperiment")
eselist <- ExploratorySummarizedExperimentList(ese)

# Almost certainly used via application creation

if (interactive()) {
  heatmap("heatmap", eselist, type = "pca")
  app <- prepare_app("heatmap", eselist)
  shiny::shinyApp(ui = app$ui, server = app$server)
}

}
\keyword{shiny}
