% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{heatmapLayout}
\alias{heatmapLayout}
\title{Produce the controls and output for a heatmap using the heatmap module}
\usage{
heatmapLayout(se, params)
}
\arguments{
\item{params}{A list object of parameters}
}
\description{
Just an abstraction to make prepareApp more concise
}
\examples{
tabPanel("Home", heatmapLayout(params))))
}
\keyword{shiny}

