% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{heatmapModuleCall}
\alias{heatmapModuleCall}
\title{Run the call to the heatmap module's server function}
\usage{
heatmapModuleCall(se, params)
}
\arguments{
\item{params}{A list object of parameters}
}
\description{
Just an abstraction to make prepareApp more concise
}
\keyword{shiny}

