% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectmatrix.R
\name{id_to_label}
\alias{id_to_label}
\title{Create row labels based on the settings of \code{labelfield} in the
\code{ExploratorySummarizedExperiment} object and the annotation data in
\code{mcols}.}
\usage{
id_to_label(ids, ese, sep = " / ")
}
\arguments{
\item{ids}{list of ids}

\item{ese}{An ExploratorySummarizedExperiment}

\item{sep}{Separator for ID and label fields}
}
\value{
String vector of same length as \code{ids}
}
\description{
Create row labels based on the settings of \code{labelfield} in the
\code{ExploratorySummarizedExperiment} object and the annotation data in
\code{mcols}.
}
\examples{
expression <- matrix(1:12, nrow = 3,
  dimnames = list(c("ENSG1", "ENSG2", "ENSG3"), paste0("s", 1:4)))
coldata <- data.frame(
  condition = rep(c("treated", "control"), each = 2),
  row.names = paste0("s", 1:4)
)
annotation <- data.frame(
  gene_id = c("ENSG1", "ENSG2", "ENSG3"),
  gene_name = c("GeneA", "GeneB", "GeneC"),
  row.names = c("ENSG1", "ENSG2", "ENSG3")
)
ese <- ExploratorySummarizedExperiment(
  assays = list(expression = expression),
  colData = coldata,
  annotation = annotation,
  idfield = "gene_id",
  labelfield = "gene_name"
)
id_to_label(c("ENSG1", "ENSG2"), ese)

}
