% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectmatrix.R
\name{idToLabel}
\alias{idToLabel}
\title{Create row labels based on the settings of \code{labelfield} in the 
SummarizedExperiment object and the annotation data in \code{mcols}}
\usage{
idToLabel(ids, ese)
}
\arguments{
\item{ese}{An ExploratorySummarizedExperiment}

\item{list}{of ids}
}
\value{
String vector of same length as \code{ids}
}
\description{
Create row labels based on the settings of \code{labelfield} in the 
SummarizedExperiment object and the annotation data in \code{mcols}
}

