% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illuminaarray.R
\name{illuminaarrayInput}
\alias{illuminaarrayInput}
\title{The input function of the illuminaarray module}
\usage{
illuminaarrayInput(id, eselist)
}
\arguments{
\item{id}{Submodule namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\value{
output An HTML tag object that can be rendered as HTML using
  as.character()
}
\description{
This provides the form elements to control the Illumina expression array
display. The illuminaarray module combines output from analysis modules such
as PCA and boxplots into a comprehensive application.
}
\examples{
illuminaarrayInput("illuminaarray", eselist)

}
\keyword{shiny}
