% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illuminaarrayqc.R
\name{illuminaarrayqc}
\alias{illuminaarrayqc}
\title{The server function of the illuminaarrayqc module}
\usage{
illuminaarrayqc(id, eselist)
}
\arguments{
\item{id}{Module namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\description{
This module plots control probes from an illumina microarray experiment.
}
\details{
This function is called directly, using the same id as its UI counterpart,
and wraps its logic in \code{moduleServer()} (see example).
}
\examples{
illuminaarayqc("myid", eselist)

}
\keyword{shiny}
