% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illuminaarrayqc.R
\name{illuminaarrayqcInput}
\alias{illuminaarrayqcInput}
\title{The input function of the illuminaarrayqc module}
\usage{
illuminaarrayqcInput(id, eselist)
}
\arguments{
\item{id}{Submodule namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\value{
output An HTML tag object that can be rendered as HTML using 
  as.character()
}
\description{
This module plots control probes from an illumina microarray experiment.
}
\details{
This funcion provides the form elements to control the display
}
\examples{
library(shinyngs)
illuminaarrayqcInput('myid', eselist)
}
\keyword{shiny}
