% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-fields.R
\name{inline_field}
\alias{inline_field}
\title{Wrap a Shiny input so its label is displayed inline}
\usage{
inline_field(field_def, label, labelwidth = 6, tooltip = NULL)
}
\arguments{
\item{field_def}{A field definition with NULL set for the label property}

\item{label}{Field label}

\item{labelwidth}{With (in units out of 12) for label}

\item{tooltip}{Optional tooltip text explaining the field, shown via a help
icon next to the label (see \code{\link{with_help_icon}})}
}
\value{
output A UI definition that can be passed to the shinyUI function.
}
\description{
Wrap a Shiny input so its label is displayed inline
}
\examples{
inline_field(numericInput("foo", label = NULL, min = 0, max = 100, value = 50), "FOO")

}
