% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlinehelp.R
\name{inlinehelpInput}
\alias{inlinehelpInput}
\title{The input function for the \code{inlinehelp} module}
\usage{
inlinehelpInput(id)
}
\arguments{
\item{id}{Modal ID. Must match that passed to \code{inlinehelpOutput}}
}
\description{
This module uses modals from \code{shinyBS} to create help text for the
current panel which displays when a help link is clicked. Input and output
functions are placed in the output function of the calling module (see
example).
}
\examples{
dendroOutput <- function(id) {
  ns <- NS(id)
  list(
    inlinehelpInput(ns('dendro')),
    inlinehelpOutput(ns('dendro'), 'dendro', 'Sample clustering dendrograms'),
    h3("Sample clustering dendrogram"), 
    plotOutput(ns("sampleDendroPlot"), height = 600)
  )
}
}

