% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-matrix.R
\name{interleaveColumns}
\alias{interleaveColumns}
\title{Interleave the columns of two matrices of equal dimensions}
\usage{
interleaveColumns(mat1, mat2)
}
\arguments{
\item{mat1}{First numeric matrix}

\item{mat2}{Second numeric matrix}
}
\value{
output Interleaved matrix
}
\description{
Interleave the columns of two matrices of equal dimensions
}
\examples{
mat1 <- matrix(1:4, nrow = 2, dimnames = list(NULL, c("a1", "a2")))
mat2 <- matrix(5:8, nrow = 2, dimnames = list(NULL, c("b1", "b2")))
interleaveColumns(mat1, mat2)

}
