% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-string.R
\name{is_valid_positive_integer_vector}
\alias{is_valid_positive_integer_vector}
\title{Check if a comma-separated string can be parsed to an integer vector}
\usage{
is_valid_positive_integer_vector(string)
}
\arguments{
\item{string}{Input string}
}
\value{
output Boolean indicating whether the check passed
}
\description{
Check if a comma-separated string can be parsed to an integer vector
}
\examples{
is_valid_positive_integer_vector("1,2,3")
is_valid_positive_integer_vector("1,two,3")

}
