% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectmatrix.R
\name{labelMatrix}
\alias{labelMatrix}
\title{Add columns to display ID and label in a table}
\usage{
labelMatrix(matrix, ese, idcol = NULL)
}
\arguments{
\item{matrix}{The input table}

\item{ese}{An ExploratorySummarizedExperiment object}
}
\value{
output Table with columns added
}
\description{
Labels only added if \code{labelfield} is specified in \code{ese}
}

