% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelselectfield.R
\name{labelselectfield}
\alias{labelselectfield}
\title{The server function of the \code{labelselectfield} module}
\usage{
labelselectfield(
  input,
  output,
  session,
  eselist,
  getExperiment = NULL,
  labels_from_all_experiments = FALSE,
  url_field = "label",
  max_items = 1,
  field_selection = FALSE,
  id_selection = FALSE,
  getNonEmptyRows = NULL,
  list_input = FALSE
)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}

\item{getExperiment}{Reactive supplying
an \code{ExploratorySummarizedExperiment}}

\item{labels_from_all_experiments}{Derive valid labels from all experiments?}

\item{url_field}{Parameter to extract from the URL to set the value of the
select field}

\item{max_items}{Maximum number of metadata values that can be selected}

\item{field_selection}{Allow selection of the meta field to use (TRUE), or
use labels (or row ids if the label field is not set)?}

\item{id_selection}{Allow users to pick specific ID from those that relate
to the specified label? (default: FALSE)}

\item{getNonEmptyRows}{Optional reactive providing non empty matrix rows}

\item{list_input}{Boolean: will input be a list of values?}
}
\description{
This module provides an input which allows filtering on the basis of data in
the metadata slot of an \code{ExploratorySummarizedExperiment}. It will only
be called by other modules requiring that input.
}
\details{
Where metadata is present, the user can select which field to select on, and
the value of that field (populated conditionall on field selction). If
specified, checkboxes are provided to allow selection of specific row IDs.

A \code{\link[shiny]{selectizeInput}} is used for performance reasons,
providing an autocomplete field for selecting from a list that could stretch
to thousands of entries. This would be difficult to do client-side using a
standard select field.
}
\examples{
callModule(labelselectfield, "myid", eselist)

}
