% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{mad_score}
\alias{mad_score}
\title{Calculate MAD scores as per OmicSoft}
\usage{
mad_score(matrix, sample_sheet = NULL, groupby = NULL, outlier_threshold = -5)
}
\arguments{
\item{matrix}{Matrix with samples by column}

\item{sample_sheet}{Sample sheet with samples by row}

\item{groupby}{Sample sheet column that can be used to group samples}

\item{outlier_threshold}{Value below which points should be flagged as
outliers, conventionally -5}
}
\value{
mad_score A data frame with columns for group name, MAD score and outlier status. A
  threshold of < -5 usually indicates outliers
}
\description{
Folllows description at
https://wiki.arrayserver.com/wiki/index.php?title=CorrelationQC.pdf.
}
\details{
Not currently deployed anywhere in shinyngs, but a potential way of flagging
outliers for investigation
}
\examples{
mat <- matrix(rnorm(60), nrow = 6, ncol = 10,
  dimnames = list(paste0("gene", 1:6), paste0("s", 1:10)))
sample_sheet <- data.frame(
  condition = rep(c("treated", "control"), each = 5),
  row.names = colnames(mat)
)
mad_score(mat, sample_sheet, groupby = "condition")

}
