% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{madScore}
\alias{madScore}
\title{Calculate MAD scores as per OmicSoft}
\usage{
madScore(matrix, sample_sheet = NULL, groupby = NULL, outlier_threshold = -5)
}
\arguments{
\item{matrix}{Matrix with samples by column}

\item{groupby}{Sample sheet column that can be used to group samples}

\item{outlier_threshold}{Value below which points should be flagged as
outliers, conventionally -5}

\item{sample.sheet}{Sample sheet with samples by row}
}
\value{
mad_score A data frame with columns for group name, MAD score and outlier status. A
  threshold of < -5 usually indicates outliers
}
\description{
Folllows description at
https://wiki.arrayserver.com/wiki/index.php?title=CorrelationQC.pdf.
}
\details{
Not currently deployed anywhere in shinyngs, but a potential way of flagging
outliers for investigation
}
