% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-from-mae.R
\name{Make_ESE_list}
\alias{Make_ESE_list}
\title{Creates a ExploratorySummarizedExperiment list object}
\usage{
Make_ESE_list(MAE_Object)
}
\arguments{
\item{MAE_Object}{List containing - minimally - a SummarizedExperiment object, species data derived from and type of transcriptomics analysis}
}
\description{
Contain one or more ExploratorysummarizedExperiments with the same sets of samples/columns but different feature sets. This facilitates the examination of expression at, for example, both transcript and gene levels or for different gene biotypes in RNA-seq experiments explorted via ‘Shinyngs’
}
