% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{makeAnnotationColors}
\alias{makeAnnotationColors}
\title{Make color sets to use in heatmap annotation}
\usage{
makeAnnotationColors(sample_annotation)
}
\arguments{
\item{sample_annotation}{A data frame with sample metadata}
}
\value{
output A list object with colors
}
\description{
Given a data frame of experimental variables, generate sets of colors, one
for each column, than can be used as the 'annotation_colors' argment to
\code{pheatmap()}. Uses \code{RColorBrewer}.
}
\examples{
colors <- makeAnnotationColors(sample_annotation)

}
