% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{makeContrastFilterSet}
\alias{makeContrastFilterSet}
\title{Make a complete set of filters for a contrast: the contrast itself, fold
change, and where applicable p- and q- values.}
\usage{
makeContrastFilterSet(
  ns,
  ese,
  assay,
  contrasts,
  contrast_numbers,
  multiple,
  show_controls,
  default_foldchange = 2,
  default_pval = 0.05,
  default_qval = 0.1,
  index = "",
  filter_rows = TRUE,
  select_all_contrasts = FALSE,
  initial_values = NULL
)
}
\arguments{
\item{ns}{A namespace function (created with \code{\link[shiny]{NS}} to be
used in creating field IDs.}

\item{ese}{ExploratorySummarizedExperiment object}

\item{assay}{Assay in \code{ese}}

\item{contrasts}{A list of lists specifying contrasts.}

\item{contrast_numbers}{A named vector of indices corresponding to
\code{contrasts}.}

\item{multiple}{Allow multiple contrasts to be selected? Passed to
\code{\link{makeContrastControl}}.}

\item{show_controls}{Show controls? Setting to false will cause them to be
hidden.}

\item{default_foldchange}{Default value for the fold change field}

\item{default_pval}{Default value for the p value field}

\item{default_qval}{Default value for the q value field}

\item{index}{Index. Will be used to differentiate mutiple copies of the
field set.}

\item{filter_rows}{Use fold change and p value etc to filter values?}

\item{select_all_contrasts}{Select all contrasts by default?}

\item{initial_values}{Optional named list of initial field values}
}
\value{
output An HTML tag object that can be rendered as HTML using
as.character()
}
\description{
Make a complete set of filters for a contrast: the contrast itself, fold
change, and where applicable p- and q- values.
}
