% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{makeHeatmap}
\alias{makeHeatmap}
\title{Make the heatmap}
\usage{
makeHeatmap(input, exprmatrix, experiment, ...)
}
\arguments{
\item{input}{Input object}

\item{...}{Arguments that should be passed through \code{annotatedHeatmap()}
to \code{pheatmap()}}

\item{se}{StructuredExperiment object with assay and experimental data}

\item{assay}{The name or index of the assay in \code{se} that should be used}
}
\value{
output Heatmap plot
}
\description{
Just an adapter to annotatedHeatmap(), which is designed to be generically
useful
}
\examples{
makeHeatmap(input, se, main = 'title')
}

