% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{makePackageCitation}
\alias{makePackageCitation}
\title{Return a usable citation string for a package}
\usage{
makePackageCitation(package)
}
\arguments{
\item{package}{Character string of package name}
}
\value{
citation A string that can be put in markdown doucments etc.
}
\description{
Adds just the tiny bit of wrapping code to make the output of
\code{citation()} printable.
}
