% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{makeUpsetPlot}
\alias{makeUpsetPlot}
\title{Make an UpSet plot using the \code{\link[UpSetR]{upset}} tool of Lex, Gehlenborg et al.}
\usage{
makeUpsetPlot(list_input, nsets = 10, nintersects = 20,
  empty.intersections = FALSE, text.scale = 1.8, point.size = 3,
  group_by = "degree")
}
\arguments{
\item{list_input}{Feature sets as a list of vectors.}

\item{nsets}{Number of sets to use}

\item{nintersects}{Number of intersections to display}

\item{empty.intersections}{Show empty intersections?}

\item{text.scale}{Scaling factor for text elements}

\item{point.size}{Point size for matrix}

\item{group_by}{Group by 'degree' or 'sets'}
}
\description{
Right now this function is a bit superfluous, passing parameters directly on
to upset.
}
\references{
Lex and Gehlenborg (2014). Points of view: Sets and intersections. <em>Nature Methods</em> 11, 779 (2014). \url{http://www.nature.com/nmeth/journal/v11/n8/abs/nmeth.3033.html} 

Gehlenborg N (2016). <em>UpSetR: A More Scalable Alternative to Venn and Euler Diagrams for Visualizing Intersecting Sets</em>. R package version 1.3.0, \url{https://CRAN.R-project.org/package=UpSetR}
}

