% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maplot.R
\name{maplot}
\alias{maplot}
\title{The server function of the \code{maplot} module}
\usage{
maplot(input, output, session, eselist)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\description{
This module is for making scatter plots comparing pairs of groups defined in
a 'tests' slot of the ExploratorySummarizedExperiment
}
\details{
This function is not called directly, but rather via callModule() (see 
example).
}
\examples{
callModule(maplot, 'maplot', eselist)
}
\keyword{shiny}

