% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maplot.R
\name{maplotInput}
\alias{maplotInput}
\title{The UI input function of the \code{maplot} module}
\usage{
maplotInput(id, eselist)
}
\arguments{
\item{id}{Submodule namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\value{
output An HTML tag object that can be rendered as HTML using
  as.character()
}
\description{
This module produces an MA plot of log(10) expression vs log(2) fold change
for contrasts defined in the `contrasts` slot of an
'ExploratorySummarizedExperimentList` object.
}
\details{
Leverages the \code{contrasts} and \code{scatterplot} modules
}
\examples{
data(airway, package = "airway")
ese <- as(airway, "ExploratorySummarizedExperiment")
eselist <- ExploratorySummarizedExperimentList(ese)

maplotInput("myid", eselist)

# Almost certainly used via application creation

if (interactive()) {
  app <- prepareApp("maplot", eselist)
  shiny::shinyApp(ui = app$ui, server = app$server)
}

}
\keyword{shiny}
