% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-from-mae.R
\name{Merge_DEA_Statistics}
\alias{Merge_DEA_Statistics}
\title{Add DEA results ExploratorySummarizedExperiment list}
\usage{
Merge_DEA_Statistics(ESEL_Object, Expt_Type, Contrasts, ESEL.Info, Phenos)
}
\arguments{
\item{ESEL_Object}{List containing - minimally - a SummarizedExperiment object, species data derived from and type of transcriptomics analysis}
}
\description{
Add DEA results ExploratorySummarizedExperiment list
}
