% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modalInput}
\alias{modalInput}
\title{The input function for the \code{modal} module}
\usage{
modalInput(
  id,
  label,
  class,
  icon = "info-circle",
  tooltip = "Open detailed help"
)
}
\arguments{
\item{id}{Modal ID. Must match that passed to \code{modalServer}}

\item{label}{A label to use for the link}

\item{class}{A class to apply to the link}

\item{icon}{Icon used to activate modal}

\item{tooltip}{Hover/focus hint describing where the link leads, shown via
\code{bslib::tooltip()}}
}
\description{
This module uses Shiny's \code{modalDialog()} to create overlaid text for the
current panel which displays when a link is clicked. The link is placed in
the output function of the calling module, and the modal is shown by
\code{modalServer()} in the calling module's server function (see example).
}
\details{
This is handy, for example when adding help text.
}
\examples{
modalInput(ns("dendro"), "help")

}
