% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modalOutput}
\alias{modalOutput}
\title{The output function of the \code{modal} module}
\usage{
modalOutput(id, title, content)
}
\arguments{
\item{id}{Modal ID. Must match that passed to \code{inlinehelpInput}}

\item{title}{Title to show on the help modal}

\item{prefix}{Prefix for the markdown file in \code{shinyngs} to display}
}
\description{
This module uses modals from \code{shinyBS} to create overlaid text for the
current panel which displays when a link is clicked. Input and output
functions are placed in the output function of the calling module (see
example).
}
\details{
This is handy, for example when adding help text.
}
\examples{
modalOutput(ns('dendro'), 'Sample clustering dendrogram', includeMarkdown(system.file('inlinehelp', 'dendro.Rmd', package = packageName())))
}
