% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{na.replace}
\alias{na.replace}
\title{Replace NAs with a string for convenience}
\usage{
na.replace(vec, replacement = "NA")
}
\arguments{
\item{vec}{Character vector or factor containing NAs}

\item{replacment}{Character replacement (default: 'NA')}
}
\value{
Vector or factor with NAs replaced
}
\description{
Replace NAs with a string for convenience
}

