% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-string.R
\name{na_replace}
\alias{na_replace}
\title{Replace NAs with a string for convenience}
\usage{
na_replace(vec, replacement = "NA")
}
\arguments{
\item{vec}{Character vector or factor containing NAs}

\item{replacement}{Character replacement (default: 'NA')}
}
\value{
Vector or factor with NAs replaced
}
\description{
Replace NAs with a string for convenience
}
\examples{
na_replace(c("a", NA, "b"))

}
