% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{nlines}
\alias{nlines}
\title{Count the number of lines in a string}
\usage{
nlines(string)
}
\arguments{
\item{string}{A string}
}
\value{
output Integer supplying the number of lines in a string
}
\description{
Count the number of lines in a string
}
\examples{
nlines('foo\nbar')
}
