% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{The server function of the pca module}
\usage{
pca(id, eselist)
}
\arguments{
\item{id}{Module namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\description{
This module calculates a PCA and formats components and loadings for display.
It uses a common set of controls, generated with the
\code{scatterplotcontrols} module, to power scatter plots for both components
and loadings produced by the \code{scatterplots} module.
}
\details{
This function is called directly, using the same id as its UI counterpart,
and wraps its logic in \code{moduleServer()} (see example).

Matrix and UI selection elements provided by the selectmatrix module
}
\examples{
data(airway, package = "airway")
ese <- as(airway, "ExploratorySummarizedExperiment")
eselist <- ExploratorySummarizedExperimentList(ese)

# Almost certainly used via application creation

if (interactive()) {
  pca("pca", eselist)
  app <- prepare_app("pca", eselist)
  shiny::shinyApp(ui = app$ui, server = app$server)
}

}
\keyword{shiny}
