% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pcaInput}
\alias{pcaInput}
\title{The input function of the pca module}
\usage{
pcaInput(id, eselist)
}
\arguments{
\item{id}{Submodule namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\value{
output An HTML tag object that can be rendered as HTML using
  as.character()
}
\description{
This provides the form elements to control the pca display, derived from the
\code{selectmatrix}, \code{scatterplotcontrols}, and \code{scatterplot}
modules.
}
\examples{
data(airway, package = "airway")
ese <- as(airway, "ExploratorySummarizedExperiment")
eselist <- ExploratorySummarizedExperimentList(ese)

pcaInput("pca", eselist)

# Almost certainly used via application creation

if (interactive()) {
  app <- prepareApp("pca", eselist)
  shiny::shinyApp(ui = app$ui, server = app$server)
}

}
\keyword{shiny}
