% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{pcaLayout}
\alias{pcaLayout}
\title{Produce the controls and output for a 3D PCA plot using the pca module, in a
shiny \code{sideBarLayout()}.}
\usage{
pcaLayout(se, params)
}
\arguments{
\item{params}{A list object of parameters}
}
\description{
Just an abstraction to make prepareApp more concise
}
\examples{
tabPanel('Home', pcaLayout(se, params))))
}
\keyword{shiny}

