% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{pcaModuleCall}
\alias{pcaModuleCall}
\title{Run the call to the pca module's server function}
\usage{
pcaModuleCall(se, params)
}
\arguments{
\item{se}{A SummarizedExperiment object containing assay data (expression, 
counts...), sample data and annotation data for the rows.}

\item{params}{A list object of parameters}
}
\description{
Just an abstraction to make prepareApp more concise
}
\keyword{shiny}

