% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pcaOutput}
\alias{pcaOutput}
\title{The output function of the pca module}
\usage{
pcaOutput(id)
}
\arguments{
\item{id}{Module namespace}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
This provides a shiny \code{tabsetPanel} with \code{tabPanel}s for both 
components and loading plots.
}
\examples{
pcaOutput('pca')

# Almost certainly used via application creation

data(zhangneurons)
app <- prepareApp('pca', zhangneurons)
shiny::shinyApp(ui = app$ui, server = app$server)
}
\keyword{shiny}

