% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdownload.R
\name{plotdownload}
\alias{plotdownload}
\title{The server function of the gene set module}
\usage{
plotdownload(id, makePlot, filename = "plot.png", plotHeight, plotWidth)
}
\arguments{
\item{id}{Module namespace}

\item{makePlot}{A reactive for generating the plot}

\item{filename}{A filename (default = 'plot.png')}

\item{plotHeight}{A number or reactive for calculating the height of the plot}

\item{plotWidth}{A number or reactive for calculating the width of the plot}
}
\description{
The plotdownload module provides export functionality for panels with plots.
This will generally not be called directly, but by other modules
}
\examples{
plotdownload("heatmap", makePlot = plotHeatmap, filename = "heatmap.png", plotHeight = plotHeight, plotWidth = plotWidth)

}
\keyword{shiny}
