% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdownload.R
\name{plotdownloadInput}
\alias{plotdownloadInput}
\title{The input function of the gene plotdownload module}
\usage{
plotdownloadInput(id, label = "Plot")
}
\arguments{
\item{id}{Submodule namespace}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
The plotdownload module provides export functionality for panels with plots.
This will generally not be called directly, but by other modules
}
\examples{
plotdownloadInput(ns('heatmap'))
}
\keyword{shiny}

