% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{plotly_quartiles}
\alias{plotly_quartiles}
\title{Make a line-based alternative to boxplots}
\usage{
plotly_quartiles(
  matrix,
  ese,
  expressiontype = "expression",
  whisker_distance = 1.5
)
}
\arguments{
\item{matrix}{Numeric matrix}

\item{ese}{ExploratorySummarizedExperiment}

\item{expressiontype}{Y axis label}

\item{whisker_distance}{IQR multiplier for whiskers, and beyond which to
show outliers (see \code{coef} in \code{\link[ggplot2]{geom_boxplot}})}
}
\description{
Box-plots become unmanagable with large numbers of samples. This function
plots lines at the median, quartiles, and whiskers, plotting points for
outliers beyond that
}
\examples{
data(airway, package = "airway")
plotly_quartiles(assays(airway)[[1]], as(airway, "ExploratorySummarizedExperiment"))

}
