% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{plotly_scatterplot}
\alias{plotly_scatterplot}
\title{Make scatterplots with \code{plot_ly()}}
\usage{
plotly_scatterplot(
  x,
  y,
  z = NULL,
  colorby = NULL,
  plot_type = "scatter",
  title = "",
  legend_title = "",
  xlab = "x",
  ylab = "y",
  zlab = "z",
  palette = NULL,
  point_size = 5,
  labels = NULL,
  show_labels = FALSE,
  lines = NULL,
  hline_thresholds = NULL,
  vline_thresholds = NULL,
  showlegend = TRUE
)
}
\arguments{
\item{x}{X coordinates}

\item{y}{Y coordinates}

\item{z}{Optional Z coordinates}

\item{colorby}{String vector or factor specifying value groups}

\item{plot_type}{Plot type: 'scatter' or 'scatter3d'}

\item{title}{Plot title}

\item{legend_title}{Legend title}

\item{xlab}{X label}

\item{ylab}{Y label}

\item{zlab}{Z label}

\item{palette}{Color palette correct for the number of groups in 'colorby'}

\item{point_size}{Main point size}

\item{labels}{Point labels}

\item{show_labels}{Permanently show labels for labelled points (default is just on hoverover)}

\item{lines}{3 column data-frame (name, x, y) with two rows, one for the
start and end of each named line}

\item{hline_thresholds}{Named list of horizontal lines with y coordinates}

\item{vline_thresholds}{Named list of vertical lines x coordinates}

\item{showlegend}{Boolean: show a legend?}
}
\value{
output Plotly plot object
}
\description{
Make scatterplots with \code{plot_ly()}
}
