% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plotlyPCA}
\alias{plotlyPCA}
\title{Make the PCA plot using plotly}
\usage{
plotlyPCA(pcavals, pcX, pcY, pcZ, pcameta, colorby = NULL, title = "foo")
}
\arguments{
\item{pcavals}{Matrix of values}

\item{pcameta}{Data frame containing metadata to use for coloring etc}

\item{colorby}{Column name in \code{pcameta} specifying how points should be colored}

\item{title}{String to use as title in plot}

\item{pcaX}{Integer value between 1 and 10 specifying which component to plot on the X axis}

\item{pcaX}{Integer value between 1 and 10 specifying which component to plot on the Y axis}

\item{pcaZ}{Integer value between 1 and 10 specifying which component to plot on the Z axis}
}
\value{
output Plotly plot object
}
\description{
Calculates a principal components analysis and plots it in 3D using the 
plotly library. Takes a few seconds for plotly to do its thing.
}
\examples{
plotlyPCA(
  mymatrix, 
  1,
  2,
  3,
  pcameta=myexperiment, 
  colorby=myvariable,
  title='My title'
) 
}

