% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesetselect.R
\name{prettify_gene_set_name}
\alias{prettify_gene_set_name}
\title{Prettify gene set names like those from MSigDB}
\usage{
prettify_gene_set_name(gsn)
}
\arguments{
\item{gsn}{Gene set name like 'KEGG_GLYCOLYSIS_GLUCONEOGENESIS'}
}
\value{
output Prettified version
}
\description{
Prettify gene set names like those from MSigDB
}
\examples{
prettify_gene_set_name("KEGG_GLYCOLYSIS_GLUCONEOGENESIS")

}
