% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-string.R
\name{prettify_variable_name}
\alias{prettify_variable_name}
\title{Make machine variable names pretty for display}
\usage{
prettify_variable_name(vn, tolower = FALSE)
}
\arguments{
\item{vn}{A string to be prettified}

\item{tolower}{Convert to lower case first? (Default: FALSE)}
}
\value{
output Prettified string
}
\description{
Just split out words using '_' and capitalise first letter
}
\examples{
vn <- "ugly_name_of_thing"
prettyifyVariablename(vn)

}
