% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\name{prettifyGeneSetName}
\alias{prettifyGeneSetName}
\title{Prettify gene set names like those from MSigDB}
\usage{
prettifyGeneSetName(gsn)
}
\arguments{
\item{gsn}{Gene set name like 'KEGG_GLYCOLOYSIS_GLUCONEOGENESIS'}
}
\value{
output Prettified version
}
\description{
Prettify gene set names like those from MSigDB
}

