% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-helpers.R
\name{push_to_list}
\alias{push_to_list}
\title{Simple list push}
\usage{
push_to_list(input_list, element)
}
\arguments{
\item{input_list}{A list to push to}

\item{element}{The element to push}
}
\value{
list with element pushed
}
\description{
Simple list push
}
\examples{
mylist <- push_to_list(mylist, "new element")

}
