% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-from-mae.R
\name{Read.MAE_RDS}
\alias{Read.MAE_RDS}
\title{Load a MultiAssayExperiment object}
\usage{
Read.MAE_RDS(rds.file)
}
\arguments{
\item{rds.file}{String - location of RDS file of the MAE object}
}
\value{
A list with a \code{se} component (the \code{SummarizedExperiment} object), and optionally \code{BioTypes}, \code{Contrast.Names} and \code{Statistics} if DEA/GSEA results are present; these contain the biotypes of the genes analysed, the contrasts and the names of the summary statistics of the DEA and GSEA runs, respectively.
}
\description{
Reads a RDS file containing the MAE object, and returns a list containing a \code{SummarizedExperiment} object with the results from differential expression & geneset enrichment analyses if performed.
}
\examples{
\dontrun{
# MultiAssayExperiment with DEA & GSE data
rds.file1 <- "path_to_rds_file1/mae-dea.rds"
dea.lst <- Read.MAE_RDS(rds.file1)

# Exploratory ShinyNGS app. - MultiAssayExperiment with only raw counts, VSTs etc.
rds.file2 <- "path_to_rds_file2/mae-explr.rds"
explr.lst <- Read.MAE_RDS(rds.file2)
}

}
