% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{read_contrasts}
\alias{read_contrasts}
\title{Read and validate a contrasts file against sample metadata}
\usage{
read_contrasts(
  filename,
  samples,
  variable_column = "variable",
  reference_column = "reference",
  target_column = "target",
  blocking_column = "blocking",
  convert_to_list = FALSE
)
}
\arguments{
\item{filename}{Contrasts file}

\item{samples}{Data frame of sample information}

\item{variable_column}{Column in contrasts file referencing sample sheet
column}

\item{reference_column}{Column in contrast file referencing reference level
of sample sheet variable}

\item{target_column}{Column in contrast file referencing target level of
sample sheet variable}

\item{blocking_column}{Colon-separated column in contrast file referencing
sample sheet variables to be used as blocking factors}

\item{convert_to_list}{Convert output to a list as used internally by
shinyngs?}
}
\value{
output Validated contrasts data frame
}
\description{
Checks: 
1. No duplicate contrast IDs. Ensure that the required columns (variable, reference, target) are present.  
2. Values in the contrast variable column exist as column names in the sample metadata.
3. If blocking factors are supplied, checks that they are present in the sample metadata.
4. Design matrix is full rank.
5. Warn about continuous covariates (e.g. numeric patient IDs treated as continuous).
6. Values of specified columns don't contain special characters.
7. Verify that the specified reference and target values exist in the corresponding sample metadata column.  
8. Issue a warning if the reference and target levels are identical.
}
