% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-read.R
\name{read_matrix}
\alias{read_matrix}
\title{Read an expression matrix file and match to specified samples and features}
\usage{
read_matrix(
  matrix_file,
  sample_metadata,
  feature_metadata = NULL,
  sep = NULL,
  row.names = 1
)
}
\arguments{
\item{matrix_file}{Matrix file}

\item{sample_metadata}{Data frame of sample metadata}

\item{feature_metadata}{Data frame of feature metadata}

\item{sep}{Separator in matrix file}

\item{row.names}{Matrix column number or name containing feature identifiers}
}
\value{
Numeric matrix
}
\description{
Read an expression matrix file and match to specified samples and features
}
\examples{
mat <- matrix(1:12, nrow = 3,
  dimnames = list(paste0("gene", 1:3), paste0("s", 1:4)))
matrix_file <- tempfile(fileext = ".tsv")
write.table(
  data.frame(gene_id = rownames(mat), mat, check.names = FALSE),
  matrix_file, sep = "\t", row.names = FALSE, quote = FALSE
)
sample_metadata <- data.frame(
  condition = rep(c("treated", "control"), each = 2),
  row.names = paste0("s", 1:4)
)
read_matrix(matrix_file, sample_metadata)

}
