% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{read_metadata}
\alias{read_metadata}
\title{Read a metadata file}
\usage{
read_metadata(filename, id_col, sep = ",", stringsAsFactors = FALSE)
}
\arguments{
\item{filename}{File name}

\item{id_col}{Identifier column in the file}

\item{sep}{File separator}
}
\value{
output Data frame
}
\description{
Read a metadata file
}
