% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdistributionplot.R
\name{readdistributionplot}
\alias{readdistributionplot}
\title{Server function of the \code{readdistributionplot} module}
\usage{
readdistributionplot(input, output, session, eselist)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{eselist}{ExploratorySummarizedExperiment with \code{read_distribution}
slot filled}

\item{id}{Submodule namespace}
}
\description{
Display a read distribution table and plot for genomic feature counts 
(exons, introns etc) vs sample, as might be generated using RSeQC.
}

