% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdistributionplot.R
\name{readdistributionplotInput}
\alias{readdistributionplotInput}
\title{Input function of the \code{readdistributionplot} module}
\usage{
readdistributionplotInput(id, eselist)
}
\arguments{
\item{id}{Submodule namespace}

\item{eselist}{ExploratorySummarizedExperiment with \code{read_distribution}
slot filled}
}
\value{
output A list of controls that can be added to a UI definition
}
\description{
Display a read distribution table and plot for genomic feature counts 
(exons, introns etc) vs sample, as might be generated using RSeQC.
}

