% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdistributionplot.R
\name{readdistributionplotOutput}
\alias{readdistributionplotOutput}
\title{Output function of the \code{readdistributionplot} module}
\usage{
readdistributionplotOutput(id, eselist)
}
\arguments{
\item{id}{Submodule namespace}

\item{eselist}{ExploratorySummarizedExperiment with \code{read_distribution}
slot filled}
}
\value{
output A list of elements that can be included in a panel
}
\description{
Display a read distribution table and plot for genomic feature counts 
(exons, introns etc) vs sample, as might be generated using RSeQC.
}

