% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readreports.R
\name{readreports}
\alias{readreports}
\title{Server function of the \code{readreports} module}
\usage{
readreports(id, eselist)
}
\arguments{
\item{id}{Module namespace}

\item{eselist}{ExploratorySummarizedExperiment with \code{read_distribution}
slot filled}
}
\description{
Display plots and tables relating to read mapping, attrition during
analysis etc.
}
