% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readreports.R
\name{readreportsInput}
\alias{readreportsInput}
\title{Input function of the \code{readreports} module}
\usage{
readreportsInput(id, eselist)
}
\arguments{
\item{id}{Submodule namespace}

\item{eselist}{ExploratorySummarizedExperiment with \code{read_distribution}
slot filled}
}
\value{
A list of controls that can be added to a UI definition
}
\description{
Display plots and tables relating to read mapping, attrition during 
analysis etc.
}
