% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readreports.R
\name{readreportsOutput}
\alias{readreportsOutput}
\title{Output function of the \code{readreports} module}
\usage{
readreportsOutput(id, eselist)
}
\arguments{
\item{id}{Submodule namespace}

\item{eselist}{ExploratorySummarizedExperiment with \code{read_distribution}
slot filled}
}
\value{
A list of elements that can be included in a panel
}
\description{
Display plots and tables relating to read mapping, attrition during 
analysis etc.
}
