% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowmetatable.R
\name{rowmetatable}
\alias{rowmetatable}
\title{The server function of the rowmetatable module}
\usage{
rowmetatable(id, eselist)
}
\arguments{
\item{id}{Module namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\description{
This function is called directly, using the same id as its UI counterpart,
and wraps its logic in \code{moduleServer()} (see example). Essentially this just passes the results of \code{colData()}
applied to the specified SummarizedExperiment object to the
\code{simpletable} module
}
\examples{
rowmetatable("rowmetatable", eselist)

}
\keyword{shiny}
