% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowmetatable.R
\name{rowmetatableInput}
\alias{rowmetatableInput}
\title{The UI input function of the rowmetatable module}
\usage{
rowmetatableInput(id, eselist)
}
\arguments{
\item{id}{Submodule namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\value{
output An HTML tag object that can be rendered as HTML using 
  as.character()
}
\description{
This module produces a simple table of the row metadata (accessed via 
\code{mcols}) in a SummarizedExperiment object. If more than one of these 
objects were specified, a select box should appear to pick the desired one for display.
}
\details{
Leverages the \code{simpletable} module
}
\examples{
experimentableInput('experiment', eselist)
}
\keyword{shiny}

