% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowmetatable.R
\name{rowmetatableOutput}
\alias{rowmetatableOutput}
\title{The output function of the rowmetatable module}
\usage{
rowmetatableOutput(id)
}
\arguments{
\item{id}{Module namespace}
}
\value{
output An HTML tag object that can be rendered as HTML using 
  as.character()
}
\description{
This module produces a simple table of the \code{colData()} in an 
ExploratorySummarizedExperiment object. If more than one of these objects
were specified, a select box should appear to pick the desired one for
display.
}
\details{
Leverages the \code{simpletable} module
}
\examples{
rowmetatableOutput('experiment')
}
\keyword{shiny}

