% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{run_clustering}
\alias{run_clustering}
\title{Partition the rows of a matrix into clusters with clara()}
\usage{
run_clustering(matrix, k)
}
\arguments{
\item{matrix}{Matrix with the rows to be clustered (features by row)}

\item{k}{Number of clusters requested}
}
\value{
output Output of \code{cluster::clara}
}
\description{
Common function for the clustering module, validating the requested
cluster count against the number of rows available before calling
\code{\link[cluster]{clara}}.
}
\examples{
run_clustering(matrix(rnorm(40), nrow = 10), 3)

}
\keyword{shiny}
