% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleselect.R
\name{sampleselect}
\alias{sampleselect}
\title{The server function of the sampleselect module}
\usage{
sampleselect(
  id,
  eselist,
  getExperiment,
  select_samples = TRUE,
  allow_summarise = TRUE
)
}
\arguments{
\item{id}{Module namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}

\item{getExperiment}{Reactive expression that returns a
\code{ExploratorySummarizedExperiment} with assays and metadata. Usually a
result of a user selection}

\item{select_samples}{Match the UI's sample-selection setting. When false,
all samples are selected.}

\item{allow_summarise}{Boolean, show controls for matrix summarisation?}
}
\value{
output A list of reactive functions for interrogating the selected
samples/ columns.
}
\description{
This module provides controls for selecting matrix columns by sample or
group name.
}
\details{
This function is called directly, using the same id as its UI counterpart,
and wraps its logic in \code{moduleServer()} (see example).
}
\examples{
selectSamples <- sampleselect("selectmatrix", getExperiment)

}
\keyword{shiny}
