% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleselect.R
\name{sampleselectInput}
\alias{sampleselectInput}
\title{The UI input function of the sampleselect module
 
This module provides controls for selecting matrix columns by sample or 
group name.}
\usage{
sampleselectInput(id, group_vars, default_groupvar)
}
\arguments{
\item{id}{Submodule namespace}

\item{group_vars}{The variables from the structured experiment that should
be used to control sample grouping in the plot}

\item{The}{default grouping variable to use}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
This will generally not be called directly, but by other modules such as the
selectmatrix module.
}
\examples{
sampleselectInput(ns('heatmap'))
}
\keyword{shiny}

