% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplotcontrols.R
\name{scatterplotcontrols}
\alias{scatterplotcontrols}
\title{Server function for scatterplotcontrols module}
\usage{
scatterplotcontrols(
  id,
  getDatamatrix,
  x = NA,
  y = NA,
  z = NA,
  makeColors = NULL,
  default_3d = TRUE
)
}
\arguments{
\item{id}{Module namespace}

\item{getDatamatrix}{Reactive expression that returns a matrix from which
coulumn headers will be used to create axis select drop-downs. The same
reactive should be supplied to the scatterplot module}

\item{x}{A value supplied for this parameter will cause a hidden field to
be generated instead of a select, useful for scatter plots that don't need
the user to select axes (default: NA)}

\item{y}{A value supplied for this parameter will cause a hidden field to
be generated instead of a select, useful for scatter plots that don't need
the user to select axes (default: NA)}

\item{z}{A value supplied for this parameter will cause a hidden field to
be generated instead of a select, useful for scatter plots that don't need
the user to select axes (default: NA)}

\item{makeColors}{Boolean: use controls for coloring?}

\item{default_3d}{Boolean: use 3D until the matching browser input is ready?}
}
\value{
output A list of reactives for accessing input values
}
\description{
This module provides controls (2D/3D, axes etc) for scatter plots, which
may then be used by one or more instances of the scatterplot module.
}
\examples{
scatterplotcontrols_reactives <- scatterplotcontrols("pca", pcaMatrix, x = 1, y = 2) # To have fixed axes rather than user-selected

}
