% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplotOutput}
\alias{scatterplotOutput}
\title{Output function for the scatterplot module}
\usage{
scatterplotOutput(id)
}
\arguments{
\item{id}{Module namespace}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
Output function for the scatterplot module
}
\examples{
scatterplotOutput('pca')
}

