% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneselect.R
\name{select_variable_genes}
\alias{select_variable_genes}
\title{Generate an integer ordering to select the n most variable genes out of a matrix}
\usage{
select_variable_genes(ntop, matrix = NULL, row_variances = NULL)
}
\arguments{
\item{ntop}{Number of genes to select}

\item{matrix}{Matrix with genes by row and samples by column}

\item{row_variances}{Numeric vector of variances, in case a precalculated set
of values should be used}
}
\value{
output A vector of integers
}
\description{
Generate an integer ordering to select the n most variable genes out of a matrix
}
\examples{
mat <- matrix(rnorm(60), nrow = 15, ncol = 4,
  dimnames = list(paste0("gene", 1:15), paste0("s", 1:4)))
select_variable_genes(ntop = 5, matrix = mat)

}
