% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectmatrix.R
\name{selectmatrix}
\alias{selectmatrix}
\title{The server function of the selectmatrix module}
\usage{
selectmatrix(input, output, session, ses, var_n = 50, var_max = 500)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{ses}{List of StructuredExperiment objects with assay and experimental
data, with additional information in the metadata() slot}

\item{var_n}{The number of rows to select when doing so by variance. Default = 50}

\item{var_max}{The maximum umber of rows to select when doing so by variance. 
Default = 500}
}
\value{
output A list of reactive functions for fetching the derived matrix 
and making a title based on its properties.
}
\description{
This module uses the geneselect and sampleselect modules to parse controls
defining matrix, row and column and return a subsetted matrix.
}
\details{
This function is not called directly, but rather via callModule() (see 
example).
}
\examples{
selectSamples <- callModule(sampleselect, 'selectmatrix', se)
}
\keyword{shiny}

