% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectmatrix.R
\name{selectmatrixInput}
\alias{selectmatrixInput}
\title{The UI input function of the selectmarix module}
\usage{
selectmatrixInput(id, eses, require_tests = FALSE)
}
\arguments{
\item{id}{Submodule namespace}

\item{eses}{List of ExploratorySummarizedExperiment objects with assay and 
experimental data}
}
\value{
output An HTML tag object that can be rendered as HTML using 
  as.character()
}
\description{
This module uses the geneselect and sampleselect modules to parse controls 
defining matrix, row and column and return a matrix.
}
\details{
This will generally not be called directly, but by other modules such as the 
heatmap module.
}
\examples{
selectmatrixInput(ns('heatmap'), eses)
}
\keyword{shiny}

