% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneselect.R
\name{selectVariableGenes}
\alias{selectVariableGenes}
\title{Generate an integer ordering to select the n most variable genes out of a matrix}
\usage{
selectVariableGenes(ntop, matrix = NULL, row_variances = NULL)
}
\arguments{
\item{matrix}{Matrix with genes by row and samples by column}

\item{row_variances}{Numeric vector of variances, in case a precalculated set
of values should be used}

\item{n_top}{Number of genes to select}
}
\value{
output A vector of integers
}
\description{
Generate an integer ordering to select the n most variable genes out of a matrix
}
