% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-string.R
\name{simple_split}
\alias{simple_split}
\title{Convenience interface to strsplit()}
\usage{
simple_split(string, sep = ",")
}
\arguments{
\item{string}{Input string}

\item{sep}{Separator}
}
\value{
output Character vector of strings
}
\description{
Convenience interface to strsplit()
}
\examples{
simple_split("a,b,c")

}
