% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{simpleApp}
\alias{simpleApp}
\title{Produce a simple app with controls and layout for a single module, in a
shiny \code{sideBarLayout()}.}
\usage{
simpleApp(eselist, module = NULL, ui_only = FALSE, ...)
}
\arguments{
\item{eselist}{List of ExploratorySummarizedExperiment objects with assay and experimental
data}

\item{module}{Character string specifying the module to use}

\item{ui_only}{Don't add server components (for UI testing)}

\item{...}{Additional arguments passed to the module output function}
}
\description{
Internal function called by prepareApp() to make simple single-function 
apps.
}
\examples{
simpleApp(eselist, 'heatmap', 'My study name')
}
\keyword{shiny}
