% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{simpleSplit}
\alias{simpleSplit}
\title{Convenience interface to strsplit()}
\usage{
simpleSplit(string, sep = ",")
}
\arguments{
\item{string}{Input string}

\item{sep}{Separator}
}
\value{
output Character vector of strings
}
\description{
Convenience interface to strsplit()
}
