% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpletable.R
\name{simpletableInput}
\alias{simpletableInput}
\title{The UI input function of the simpletable module
 
This module produces a simple datatable output with a download button.}
\usage{
simpletableInput(id, description = NULL)
}
\arguments{
\item{id}{Submodule namespace}

\item{description}{A string to display in the side bar to explain the table.}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
The UI input function of the simpletable module
 
This module produces a simple datatable output with a download button.
}
\examples{
selectmatrixInput(ns('heatmap'), se, group_vars, default_groupvar)
}
\keyword{shiny}

