% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{simpletableLayout}
\alias{simpletableLayout}
\title{Produce the controls and output for a datatable, in a shiny
\code{sideBarLayout()}.}
\usage{
simpletableLayout(se, params)
}
\arguments{
\item{params}{A list object of parameters}
}
\description{
Just an abstraction to make prepareApp more concise
}
\examples{
tabPanel('Home', simpletableLayout(se, params))))
}
\keyword{shiny}

