% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectmatrix.R
\name{single_valid_matrix}
\alias{single_valid_matrix}
\title{Is there only one matrix to plot from this object?}
\usage{
single_valid_matrix(eselist)
}
\arguments{
\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\value{
output Logical value
}
\description{
Convenience function for deciding how to construct filters
}
\examples{
data(airway, package = "airway")
ese <- as(airway, "ExploratorySummarizedExperiment")
eselist <- ExploratorySummarizedExperimentList(ese, title = "Airway study")
single_valid_matrix(eselist)

}
