% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectmatrix.R
\name{singleValidMatrix}
\alias{singleValidMatrix}
\title{Is there only one matrix to plot from this object?}
\usage{
singleValidMatrix(eselist)
}
\arguments{
\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\value{
output Logical value
}
\description{
Convenience function for deciding how to construct filters
}
