% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-string.R
\name{split_string_to_fixed_width_lines}
\alias{split_string_to_fixed_width_lines}
\title{Given a string with spaces, try to split into multiple lines of <
\code{linewidth} characters}
\usage{
split_string_to_fixed_width_lines(string, linewidth = 20)
}
\arguments{
\item{string}{A string with spaces}

\item{linewidth}{The maximum line length in characters (default: 30)}
}
\value{
A string with newline characters added where appropriate
}
\description{
Given a string with spaces, try to split into multiple lines of <
\code{linewidth} characters
}
\examples{
split_string_to_fixed_width_lines("once upon a time there was a giant and a beanstalk and a pot of gold and some beans")

}
