% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{splitStringToFixedwidthLines}
\alias{splitStringToFixedwidthLines}
\title{Given a string with spaces, try to split into multiple lines of <
\code{linewidth} characters}
\usage{
splitStringToFixedwidthLines(string, linewidth = 20)
}
\arguments{
\item{string}{A string with spaces}

\item{linewidth}{The maximum line length in characters (default: 20)}
}
\value{
A string with newline characters added where appropriate
}
\description{
Given a string with spaces, try to split into multiple lines of <
\code{linewidth} characters
}
\examples{
splitStringToFixedwidthLines("once upon a time there was a giant and a beanstalk and a pot of gold and some beans")

}
