% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{static_scatterplot}
\alias{static_scatterplot}
\title{Make scatterplots with \code{ggplot()} or \code{scatterplot3d}}
\usage{
static_scatterplot(
  x,
  y,
  z = NULL,
  colorby = NULL,
  plot_type = "scatter",
  title = "",
  legend_title = NULL,
  xlab = "x",
  ylab = "y",
  zlab = "z",
  palette = NULL,
  point_size = 1,
  labels = colorby,
  show_labels = FALSE,
  hline_thresholds = NULL,
  vline_thresholds = NULL,
  showlegend = TRUE,
  palette_name = "Set1"
)
}
\arguments{
\item{x}{X coordinates}

\item{y}{Y coordinates}

\item{z}{Optional Z coordinates}

\item{colorby}{String vector or factor specifying value groups}

\item{plot_type}{Plot type: 'scatter' (ggplot) or 'scatter3d' (scatterplot3d)}

\item{title}{Plot title}

\item{legend_title}{Legend title}

\item{xlab}{X label}

\item{ylab}{Y label}

\item{zlab}{Z label}

\item{palette}{Color palette correct for the number of groups in 'colorby'}

\item{point_size}{Main point size}

\item{labels}{Point labels}

\item{show_labels}{Permanently show labels for labelled points}

\item{hline_thresholds}{Named list of horizontal lines with y coordinates}

\item{vline_thresholds}{Named list of vertical lines x coordinates}

\item{showlegend}{Boolean: show a legend?}

\item{palette_name}{Valid R color palette name}
}
\value{
output Ouput object from ggplot or scatterplot3d.
}
\description{
These are not used in the shinyngs UI, but are provided here to be fairly
consistent with the plotly-driven display, and provide a static alternative
for external users.
}
