% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{stringsToNamedVector}
\alias{stringsToNamedVector}
\title{Take two delimiter-separated strings and generate a named vector}
\usage{
stringsToNamedVector(
  elements_string,
  names_string = NULL,
  sep = ",",
  prettify_names = TRUE,
  simplify_files = FALSE
)
}
\arguments{
\item{elements_string}{String to be converted to vector elements}

\item{names_string}{String to be converted to vector names by default taken
from elements_string.}

\item{sep}{Separator to use}

\item{prettify_names}{Boolean. Prettify element names?}

\item{simplify_files}{If elements are file and to be used to generate names,
should we simplify by striping path and extension?}
}
\value{
output Named character vector
}
\description{
Take two delimiter-separated strings and generate a named vector
}
