% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploratorySummarizedExperimentList-class.R
\name{[,ExploratorySummarizedExperimentList,logical,missing,ANY-method}
\alias{[,ExploratorySummarizedExperimentList,logical,missing,ANY-method}
\title{Extract parts of ExploratorySummarizedExperimentList with booleans}
\usage{
\S4method{[}{ExploratorySummarizedExperimentList,logical,missing,ANY}(x, i, j, ..., drop = TRUE)
}
\description{
Extract parts of ExploratorySummarizedExperimentList with booleans
}
