% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploratorySummarizedExperimentList-class.R
\name{[,ExploratorySummarizedExperimentList,numeric,missing,ANY-method}
\alias{[,ExploratorySummarizedExperimentList,numeric,missing,ANY-method}
\title{Extract parts of ExploratorySummarizedExperimentList with integers like 1}
\usage{
\S4method{[}{ExploratorySummarizedExperimentList,numeric,missing,ANY}(x, i, j, ..., drop = TRUE)
}
\description{
Extract parts of ExploratorySummarizedExperimentList with integers like 1
}
