% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarisematrix.R
\name{summarisematrix}
\alias{summarisematrix}
\title{The server function of the summarisematrix module}
\usage{
summarisematrix(id)
}
\arguments{
\item{id}{Module namespace}
}
\description{
This module provides a form element and associated get function for defining
how a summary statistic is calculated (probably by mean).
}
\examples{
summarisematrix()

}
\keyword{shiny}
