% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarisematrix.R
\name{summarisematrixInput}
\alias{summarisematrixInput}
\title{The input function of the summarizematrix module}
\usage{
summarisematrixInput(id, allow_none = TRUE, select_summary_type = TRUE)
}
\arguments{
\item{id}{Submodule namespace}

\item{allow_none}{Allow a 'no summarisation' selection.}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
This module provides a form element and associated get function for defining
how a summary statistic is calculated (probably by mean).
}
\examples{
contrastsInput('test')
}
\keyword{shiny}

