% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarisematrix.R
\name{summarizeMatrix}
\alias{summarizeMatrix}
\title{Summarise the rows of a matrix, applying a function to groups of cells 
defined by a factor}
\usage{
summarizeMatrix(matrix, treatment_factor, summaryFunc = "colMeans")
}
\arguments{
\item{matrix}{Numeric matrix}

\item{treatment_factor}{a factor defining column groups}

\item{summaryFunc}{A Function to apply to a transformed version of 
\code{matrix} (default: colMeans)}
}
\value{
Summarized matrix, with e.g. means in columns
}
\description{
Note that the specified function will be applied to a tranformed version
of the matrix, so \code{colMeans()}, for example, is appropriate.
}
\examples{
summarizeMatrix(mymatrix, myfactor)
}

