% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{summary_se}
\alias{summary_se}
\title{Summarise an input matrix}
\usage{
summary_se(
  data = NULL,
  measurevar,
  groupvars = NULL,
  na.rm = FALSE,
  conf.interval = 0.95,
  add_medians = FALSE,
  .drop = TRUE
)
}
\arguments{
\item{data}{A data frame. Expects all values for summarisation to be in one
column, which may require judicious use of \code{melt_matrix}.}

\item{measurevar}{The name of a column that contains the variable to be
summariezed}

\item{groupvars}{A vector containing names of columns that contain grouping
variables}

\item{na.rm}{A boolean that indicates whether to ignore NA's}

\item{conf.interval}{The percent range of the confidence interval (default is
95 percent)}

\item{add_medians}{Logical indicating whether medians should be added to the
output. Standard error estimates for the median require bootstrapping, so
TRUE for this variables make summary statistic calculation take longer.}

\item{.drop}{Logical controlling whether unobserved combinations of
\code{groupvars} are dropped, passed to \code{dplyr::group_by()}.}
}
\value{
out Data frame with summary statistics
}
\description{
Matrix summarization function adapted from
\url{http://www.cookbook-r.com/Graphs/Plotting_means_and_error_bars_(ggplot2)}{the R cookbook}.
}
\details{
Uses bootstrapping to return the standard error/ deviation of the median.
}
\examples{
tg <- ToothGrowth
summary_se(tg, measurevar = "len", groupvars = c("supp", "dose"))

}
