% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topgeneboxplot.R
\name{topgeneboxplot}
\alias{topgeneboxplot}
\title{The server function of the topgeneboxplot module}
\usage{
topgeneboxplot(id, eselist)
}
\arguments{
\item{id}{Module namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\description{
This function is called directly, using the same id as its UI counterpart,
and wraps its logic in \code{moduleServer()} (see example).
}
\examples{
topgeneboxplot("topgeneboxplot", eselist)

}
\keyword{shiny}
