% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topgeneboxplot.R
\name{topgeneboxplotOutput}
\alias{topgeneboxplotOutput}
\title{The output function of the topgeneboxplot module}
\usage{
topgeneboxplotOutput(id)
}
\arguments{
\item{id}{Submodule namespace}
}
\value{
output An HTML tag object that can be rendered as HTML using
as.character()
}
\description{
The output function of the topgeneboxplot module
}
\examples{
topgeneboxplotOutput("topgeneboxplot")

}
\keyword{shiny}
