% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{ucfirst}
\alias{ucfirst}
\title{Capitalise the first letter of a string}
\usage{
ucfirst(string)
}
\arguments{
\item{string}{A string}
}
\value{
output String with capitalised first letter
}
\description{
Capitalise the first letter of a string
}
\examples{
> ucfirst('Example')
[1] "Example"
}

