% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{unpack.list}
\alias{unpack.list}
\title{Unpack a list to the environment. Handy when many reactive functions are 
returned by a call to a module's server function}
\usage{
unpack.list(object)
}
\arguments{
\item{x}{A named list of objects to unpack}
}
\description{
Unpack a list to the environment. Handy when many reactive functions are 
returned by a call to a module's server function
}

