% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{upset_filter_intersections_by_order}
\alias{upset_filter_intersections_by_order}
\title{Drop intersections involving fewer than a minimum number of sets}
\usage{
upset_filter_intersections_by_order(ints, minorder)
}
\arguments{
\item{ints}{A list with \code{combinations} and \code{intersections}, as
returned by \code{\link{upset_calculate_intersections}}}

\item{minorder}{Minimum number of sets that must be involved in an
intersection for it to be kept}
}
\value{
output \code{ints}, with both elements filtered to the same subset
}
\description{
Drop intersections involving fewer than a minimum number of sets
}
\keyword{internal}
