% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{upset_grid_plot}
\alias{upset_grid_plot}
\title{Make the grid of points indicating set membership in intersections}
\usage{
upset_grid_plot(sets, ints, nintersects)
}
\arguments{
\item{sets}{A named list of character vectors, one per set}

\item{ints}{A list with \code{combinations} and \code{intersections}, as
returned by \code{\link{upset_calculate_intersections}} (optionally
filtered by \code{\link{upset_filter_intersections_by_order}})}

\item{nintersects}{Maximum number of intersections to display, taken from
the front of \code{ints}}
}
\value{
output A plotly htmlwidget
}
\description{
Make the grid of points indicating set membership in intersections
}
\keyword{internal}
