% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{upset_intersect_size_chart}
\alias{upset_intersect_size_chart}
\title{Make the bar chart illustrating intersect size}
\usage{
upset_intersect_size_chart(ints, nintersects, bar_numbers = FALSE)
}
\arguments{
\item{ints}{A list with \code{combinations} and \code{intersections}, as
returned by \code{\link{upset_calculate_intersections}} (optionally
filtered by \code{\link{upset_filter_intersections_by_order}})}

\item{nintersects}{Maximum number of intersections to display}

\item{bar_numbers}{Add value labels above the bars?}
}
\value{
output A plotly htmlwidget
}
\description{
Make the bar chart illustrating intersect size
}
\keyword{internal}
