% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{upset_set_size_chart}
\alias{upset_set_size_chart}
\title{Make the bar chart illustrating set sizes}
\usage{
upset_set_size_chart(sets)
}
\arguments{
\item{sets}{A named list of character vectors, one per set}
}
\value{
output A plotly htmlwidget
}
\description{
Make the bar chart illustrating set sizes
}
\keyword{internal}
