% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{validate_indices}
\alias{validate_indices}
\title{Validate assay indices based on a given string.}
\usage{
validate_indices(assay_data, index_string)
}
\arguments{
\item{assay_data}{A list containing matrices as assay data.}

\item{index_string}{A string that can be a comma-separated list of integers or assay names.}
}
\value{
A vector of valid indices (either as integers or assay names).
}
\description{
This function checks if the provided index string represents valid assays in the given assay data.
The function can handle index strings that are comma-separated integers or assay names.
}
\examples{
assay_data_example <- list(a = matrix(1:9, ncol=3), b = matrix(1:12, ncol=3), c = matrix(1:6, ncol=2))
valid_assays1 <- validate_indices(assay_data_example, "1,2")
valid_assays2 <- validate_indices(assay_data_example, "a,b")

}
