% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoplot.R
\name{volcanoplot}
\alias{volcanoplot}
\title{The server function of the \code{volcanoplot} module}
\usage{
volcanoplot(id, eselist)
}
\arguments{
\item{id}{Module namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\description{
This function is called directly, using the same id as its UI counterpart,
and wraps its logic in \code{moduleServer()} (see example). Essentially this just passes the results of \code{colData()}
applied to the specified SummarizedExperiment object to the
\code{simpletable} module
}
\examples{
differentialtable("differentialtable", eselist)

# However, almost certainly called via application creation:

data(zhangneurons)
app <- prepareApp("volcanoplot", zhangneurons)
shinyApp(ui = app$ui, server = app$server)

}
\keyword{shiny}
