% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoplot.R
\name{volcanoplotOutput}
\alias{volcanoplotOutput}
\title{The output function of the differentialtable module}
\usage{
volcanoplotOutput(id)
}
\arguments{
\item{id}{Module namespace}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
This module provides information on the comparison betwen pairs of groups 
defined in a 'tests' slot of the SummarizedExperiment metadata.
}
\details{
Leverages the \code{scatterplot} module
}
\examples{
differentialtableOutput('experiment')
}
\keyword{shiny}

