% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{addTextLabels}
\alias{addTextLabels}
\title{Add permanent text labels to points in a plotly graph}
\usage{
addTextLabels(
  p,
  x,
  y,
  z,
  colorby = NULL,
  labels,
  plot_type,
  show_labels = TRUE
)
}
\arguments{
\item{p}{Previously generated plotly object}

\item{x}{Vector of numeric x values}

\item{y}{Vector of numeric y values}

\item{z}{Optional vector of numeric z values}

\item{colorby}{String vector or factor specifying value groups}

\item{labels}{Vector of labels to apply}

\item{plot_type}{Plot type: 'scatter' or 'scatter3d'}

\item{show_labels}{If false, simpy pass through input plot object}
}
\value{
output Plotly object
}
\description{
Add permanent text labels to points in a plotly graph
}
