% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{annotatedHeatmap}
\alias{annotatedHeatmap}
\title{Make a heatmap with annotations by experimental variable}
\usage{
annotatedHeatmap(plotmatrix, sample_annotation, group_vars = NULL, ...)
}
\arguments{
\item{plotmatrix}{Expression/ other data matrix}

\item{sample_annotation}{Annotation for the columns of plotmatrix}

\item{group_vars}{Grouping variables to be used from 
sample_annotation}

\item{...}{Other arguments to pass to \code{pheatmap()}}
}
\value{
output A plot as produced by pheatmap()
}
\description{
This is a generic function which may be useful outside of this package. It
produces a heatmap based on an expression matrix and accompanying 
experiment data in the form of a frame, using pheatmap().
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

