% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{annotatedHeatmap}
\alias{annotatedHeatmap}
\title{Make a heatmap with annotations by experimental variable}
\usage{
annotatedHeatmap(plotmatrix, displaymatrix, sample_annotation, cluster_cols,
  cluster_rows, scale, row_labels, row_height = 12)
}
\arguments{
\item{plotmatrix}{Expression/ other data matrix}

\item{displaymatrix}{A matrix of values that might be displayed in cells}

\item{cluster_cols}{Cluster columns?}

\item{cluster_rows}{Cluster rows?}

\item{scale}{'row', 'column' or none}

\item{row_labels}{Vector labels to use for rows}

\item{row_height}{The height to use for each row}
}
\value{
output A plot as produced by pheatmap()
}
\description{
This is a generic function which may be useful outside of this package. It
produces a heatmap based on an expression matrix and accompanying 
experiment data in the form of a frame, using \code{pheatmap()} or 
\code{d3heatmap()}.
}
\details{
The clustering parameters for \code{pheatmap()} and \code{d3heatmap()} are 
set to be consistent with one another.
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

