% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{anova_pca_metadata}
\alias{anova_pca_metadata}
\title{Generate a matrix of anova values for associating principal components with
categorical covariates.}
\usage{
anova_pca_metadata(pca_coords, pcameta, fraction_explained)
}
\arguments{
\item{pca_coords}{Data frame of PCA coordinates, with samples by row and
components by column.}

\item{pcameta}{Data frame of sample metadata with sample identifiers by row
and variables by column.}

\item{fraction_explained}{Numeric vector containing the percent contribution
to variance of each component}
}
\value{
output A numeric matrix of p values
}
\description{
Generate a matrix of anova values for associating principal components with
categorical covariates.
}
