% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assaydatatable.R
\name{assaydatatableInput}
\alias{assaydatatableInput}
\title{The UI input function of the assaydatatable module}
\usage{
assaydatatableInput(id, eselist)
}
\arguments{
\item{id}{Submodule namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\value{
output An HTML tag object that can be rendered as HTML using
  as.character()
}
\description{
This module displays the content of the currently selected experiment and
assay, also allowing grouping by mean etc.
}
\details{
Leverages the \code{simpletable} module
}
\examples{
assaydatatableInput("experiment", eselist)

# Almost certainly used via application creation

data(zhangneurons)
app <- prepareApp("assaydatatable", zhangneurons)
shiny::shinyApp(ui = app$ui, server = app$server)

}
\keyword{shiny}
