% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assaydatatable.R
\name{assaydatatableOutput}
\alias{assaydatatableOutput}
\title{The output function of the assaydatatable module}
\usage{
assaydatatableOutput(id)
}
\arguments{
\item{id}{Module namespace}
}
\value{
output An HTML tag object that can be rendered as HTML using
as.character()
}
\description{
This module displays the content of the currently selected experiment and
assay, also allowing grouping by mean etc.
}
\details{
Leverages the \code{simpletable} module
}
\examples{
assaydatatableOutput("experiment")

# Almost certainly used via application creation

data(zhangneurons)
app <- prepareApp("assaydatatable", zhangneurons)
shiny::shinyApp(ui = app$ui, server = app$server)

}
\keyword{shiny}
