% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{barplotInput}
\alias{barplotInput}
\title{Input function of the \code{barplot} module}
\usage{
barplotInput(id, default_mode = "stack", allow_select = TRUE)
}
\arguments{
\item{id}{Submodule namespace}

\item{default_mode}{Default bar mode}

\item{allow_select}{Allow user to choose stack mode?}
}
\value{
A list of controls that can be added to a UI definition
}
\description{
Generic module to isplay grouped, stacked or overlaid bars for a matrix
}
