% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{barplotOutput}
\alias{barplotOutput}
\title{Output function of the \code{barplot} module}
\usage{
barplotOutput(id, height = "400")
}
\arguments{
\item{id}{Submodule namespace}
}
\value{
A list of elements that can be included in a panel
}
\description{
Display grouped, stacked or overlaid bars for a matrix
}

