% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot}
\alias{boxplot}
\title{The server function of the boxplot module}
\usage{
boxplot(input, output, session, eselist)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\description{
This module produces displays of the quartiles of the values in the 
selected assay matrix. For low sample numbers (<= 20) the default is a 
boxplot produced using \code{ggplot2}. For higher sample numbers, the default is
a line-based alternative using \code{plotly}.
}
\details{
This function is not called directly, but rather via callModule() (see 
example).
}
\examples{
callModule(boxplot, 'boxplot', eselist)

# Almost certainly used via application creation

data(zhangneurons)
app <- prepareApp('boxplot', zhangneurons)
shiny::shinyApp(ui = app$ui, server = app$server)
}
\keyword{shiny}
