% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplotInput}
\alias{boxplotInput}
\title{The input function of the boxplot module}
\usage{
boxplotInput(id, eses)
}
\arguments{
\item{id}{Submodule namespace}

\item{eses}{List of ExploratorySummarizedExperiment objects with assay and
experimental data}
}
\value{
output An HTML tag object that can be rendered as HTML using 
  as.character()
}
\description{
This provides the form elements to control the pca display
}
\examples{
boxplotInput(ns('boxplot'), eses)
}
\keyword{shiny}

