% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplotInput}
\alias{boxplotInput}
\title{The input function of the boxplot module}
\usage{
boxplotInput(id, eselist)
}
\arguments{
\item{id}{Submodule namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\value{
output An HTML tag object that can be rendered as HTML using 
  as.character()
}
\description{
This module produces displays of the quartiles of the values in the 
selected assay matrix. For low sample numbers (<= 20) the default is a 
boxplot produced using \code{ggplot2}. For higher sample numbers, the default is
a line-based alternative using \code{plotly}.
}
\examples{
boxplotInput(ns('boxplot'), eselist)

# Almost certainly used via application creation

data(zhangneurons)
app <- prepareApp('boxplot', zhangneurons)
shiny::shinyApp(ui = app$ui, server = app$server)
}
\keyword{shiny}

