% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipseq.R
\name{chipseq}
\alias{chipseq}
\title{The server function of the chipseq module. Currently a near-clone of the
RNA-seq module, with ChIP-seq optimisations planned.}
\usage{
chipseq(id, eselist)
}
\arguments{
\item{id}{Module namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\description{
This function is called directly, using the same id as its UI counterpart,
and wraps its logic in \code{moduleServer()} (see example).
}
\examples{
chipseq("chipseq", eselist)

}
\keyword{shiny}
