% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipseq.R
\name{chipseq}
\alias{chipseq}
\title{The server function of the chipseq module. Currently a near-clone of the 
RNA-seq module, with ChIP-seq optimisations planned.}
\usage{
chipseq(input, output, session, eselist)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\description{
This function is not called directly, but rather via callModule() (see 
example).
}
\examples{
callModule(chipseq, 'chipseq', eselist)
}
\keyword{shiny}
