% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipseq.R
\name{chipseqInput}
\alias{chipseqInput}
\title{The input function of the chipseq module. Currently a near-clone of the 
RNA-seq module, with ChIP-seq optimisations planned.}
\usage{
chipseqInput(id, eselist)
}
\arguments{
\item{id}{Submodule namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\value{
output An HTML tag object that can be rendered as HTML using 
  as.character()
}
\description{
This provides the form elements to control the ChIP-seq display
}
\details{
The chipseq module is a combination of output from many modules (pca, boxplot 
etc) to form a comprehensive analysis application.
}
\examples{
chipseqInput('chipseq', eselist)
}
\keyword{shiny}
