% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro.R
\name{clusteringDendrogram}
\alias{clusteringDendrogram}
\title{Make a clustering dendrogram with coloring by experimental variable}
\usage{
clusteringDendrogram(
  plotmatrix,
  experiment,
  colorby = NULL,
  cor_method = "pearson",
  cluster_method = "ward.D",
  plot_title = "",
  labelspace = 0.2,
  palette = NULL,
  palette_name = "Set1"
)
}
\arguments{
\item{plotmatrix}{Expression/ other data matrix}

\item{experiment}{Annotation for the columns of plotmatrix}

\item{colorby}{Column name in \code{experiment} specifying how boxes should be colored}

\item{cor_method}{Correlation method, passed to cor() (default: pearson).}

\item{cluster_method}{Clustering method, passed to hclust() (default:
ward.D).}

\item{plot_title}{Plot title}

\item{labelspace}{Vertical fraction of plot to be used for labels (default: 0.2).}

\item{palette}{Palette of colors, one for each unique value derived from
\code{colorby}.}

\item{palette_name}{Valid R color palette name}
}
\value{
output A \code{ggplot} output
}
\description{
A simple function using \code{ggdendro} to make a sample dendrogram
}
\examples{
# Make a dendrogram with the data in airway

require(airway)
data(airway, pakckage = "airway")
clusteringDendrogram(assays(airway)[[1]], data.frame(colData(airway)), colorby = "dex")

# Do the same, but only usig the 1000 most variant rows and see how the
# clustering improves.

mymatrix <- assays(airway)[[1]]
mymatrix <- mymatrix[order(apply(mymatrix, 1, var), decreasing = TRUE)[1:1000], ]
clusteringDendrogram(mymatrix, data.frame(colData(airway)), colorby = "dex")

}
\keyword{keywords}
