% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarisematrix.R
\name{colGeomMeans}
\alias{colGeomMeans}
\title{Geometric means by matrix column}
\usage{
colGeomMeans(x)
}
\arguments{
\item{x}{A matrix}
}
\value{
Vector with column geometric means
}
\description{
Geometric means by matrix column
}

