% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colormaker.R
\name{colormaker}
\alias{colormaker}
\title{The output function of the colorby module}
\usage{
colormaker(id, getNumberCategories)
}
\arguments{
\item{id}{Module namespace}

\item{getNumberCategories}{A reactive supplying the number of categories
that require a color.}
}
\value{
output A reactive returning a character vector of hex colors
}
\description{
Supplies a reactive returning the palette selected in the drop-down, sized
to the number of categories the caller reports (see
\code{\link{makeColorScale}}).
}
\details{
This function is called directly, using the same id as its UI counterpart,
and wraps its logic in \code{moduleServer()} (see example).
}
\examples{
colormaker("myid", getNumberCategories)

}
\keyword{shiny}
