% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colormaker.R
\name{colormaker}
\alias{colormaker}
\title{The output function of the colorby module}
\usage{
colormaker(input, output, session, getNumberCategories)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{getNumberCategories}{A reactive supplying the number of categories 
that require a color.}
}
\value{
output An HTML tag object that can be rendered as HTML using 
  as.character()
}
\description{
This module provides a drop-down for picking an RColorBrewer color palette
and provides that palette given a reactive which supplied the required 
number of colors.
}
\details{
This function is not called directly, but rather via callModule() (see 
example).
}
\examples{
callModule(colormaker, 'myid', getNumberCategories)
}
\keyword{shiny}
