% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colormaker.R
\name{colormakerInput}
\alias{colormakerInput}
\title{The input function of the colorby module}
\usage{
colormakerInput(id)
}
\arguments{
\item{id}{Submodule namespace}
}
\value{
output NULL
}
\description{
shinyngs colours every plot from the single, fixed colour-blind-safe
palette in \code{\link{makeColorScale}}, so there is no per-plot palette
choice to expose. Modules that build their form from a list of inputs can
include this call unconditionally; it contributes no UI element.
}
\examples{
colormakerInput("myid")

}
\keyword{shiny}
