% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colormaker.R
\name{colormakerInput}
\alias{colormakerInput}
\title{The input function of the colorby module}
\usage{
colormakerInput(id)
}
\arguments{
\item{id}{Submodule namespace}
}
\value{
output An HTML tag object that can be rendered as HTML using
  as.character()
}
\description{
This module provides a drop-down for picking an RColorBrewer color palette
and provides that palette given a reactive which supplied the required
number of colors.
}
\details{
This funcion provides the form elements to control the display
}
\examples{
colormakerInput("myid")

}
\keyword{shiny}
