% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{compile_contrast_data}
\alias{compile_contrast_data}
\title{Compile contrast stats for inclusion in shinyngs}
\usage{
compile_contrast_data(
  differential_stats_files,
  sep = "\\t",
  feature_id_column = NULL,
  pval_column = NULL,
  qval_column = NULL,
  fc_column = NULL,
  unlog_folchanges = FALSE
)
}
\arguments{
\item{differential_stats_files}{Tabular files with differential stats}

\item{sep}{Separator in stats files}

\item{pval_column}{P value column in stats files}

\item{qval_column}{Q value column in stats files}

\item{fc_column}{Fold change column in stats files}

\item{feature_id_col}{Column in stats files identifiying features (e.g. genes)}
}
\value{
output A named list of data frames by statistic, number of columns equal to input file number
}
\description{
Compile contrast stats for inclusion in shinyngs
}
