% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{cond_log2_transform_assays}
\alias{cond_log2_transform_assays}
\title{Apply log2 transformation on assay data based on log2_assays parameter.}
\usage{
cond_log2_transform_assays(assay_data, log2_assays)
}
\arguments{
\item{assay_data}{A list containing assay data.}

\item{log2_assays}{A string parameter that can be NULL, empty, or a non-empty string.
If NULL: log2 transformation will be guessed based on input assays.
If empty: no log2 transformation will be applied.
If non-empty: log2 transformation will be applied to specified assays.}
}
\value{
A modified assay_data list.
}
\description{
Apply log2 transformation on assay data based on log2_assays parameter.
}
\examples{
data <- list(a = c(22,23,24), b = c(2,3,4), c = c(40,41,42))
cond_log2_transform_assays(data, NULL)
cond_log2_transform_assays(data, "")
cond_log2_transform_assays(data, "a,b")
cond_log2_transform_assays(data, "1,2")
}
