% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{contrasts}
\alias{contrasts}
\title{The server function of the contrasts module}
\usage{
contrasts(input, output, session, eselist, getExperiment = NULL,
  selectMatrix = NULL, selectColData = NULL, getAssay = NULL,
  getMetafields = NULL, multiple = FALSE, show_controls = TRUE)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{getExperiment}{Reactive for getting the selected experiment. Probably 
get this from the \code{selectmatrix} module}

\item{selectMatrix}{Reactive for generating a matrix to do comparisons with.
Probably returned by a call to the \code{\link{selectmatrix}} module.}

\item{selectColData}{Reactive returning column metadata for the matrix 
provided by selectMatrix. Probably returned by a call to the 
\code{\link{selectmatrix}} module.}

\item{getAssay}{Reactive for fetching the current assay.}

\item{multiple}{Allow selection of multiple contrasts?}

\item{show_controls}{Show the controls for contrast selection?}
}
\description{
This function is not called directly, but rather via callModule() (see 
example).
}
\examples{
callModule(contrasts, 'differential', getExperiment = getExperiment, selectMatrix = selectMatrix, getAssay = getAssay, multiple = TRUE)
}
\keyword{shiny}

