% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{contrasts}
\alias{contrasts}
\title{The server function of the contrasts module}
\usage{
contrasts(
  input,
  output,
  session,
  eselist,
  selectmatrix_reactives = list(),
  multiple = FALSE,
  select_all_contrasts = FALSE,
  show_controls = TRUE,
  default_foldchange = 2,
  default_pval = 0.05,
  default_qval = 0.1
)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}

\item{selectmatrix_reactives}{The list of reactive expressions returned by
a call to the \code{\link{selectmatrix}} module. This will be unpacked to
gain access to the data provided by those reaactive.}

\item{multiple}{Allow selection of multiple contrasts?}

\item{select_all_contrasts}{Select all contrasts by default?}

\item{show_controls}{Show the controls for contrast selection?}

\item{default_foldchange}{default value for the fold change filter}

\item{default_pval}{Default value for the p value field}

\item{default_qval}{Default value for the q value field}
}
\description{
This module provides the form elements to control contrasts used in e.g.
differential expression panels. In particular it provides the ability for
users to add filters to progressively refine a query.
}
\details{
This function is not called directly, but rather via callModule() (see
example).
}
\examples{
callModule(contrasts, "differential", getExperiment = getExperiment, selectMatrix = selectMatrix, getAssay = getAssay, multiple = TRUE)

}
\keyword{shiny}
