% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{contrastsInput}
\alias{contrastsInput}
\title{The input function of the contrasts module}
\usage{
contrastsInput(id, default_min_foldchange = 2, default_max_q = 0.1)
}
\arguments{
\item{id}{Submodule namespace}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
This module provides the form elements to control contrasts used in e.g. 
differential expression panels.
}
\examples{
contrastsInput('test')
}
\keyword{shiny}

