% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectmatrix.R
\name{convertIds}
\alias{convertIds}
\title{Convert row names to metadata identifiers}
\usage{
convertIds(ids, ese, to, remove_na = FALSE)
}
\arguments{
\item{ids}{IDs found as row names in the
\code{ExploratorySummarizedExperiment}}

\item{ese}{The \code{ExploratorySummarizedExperiment}}

\item{to}{The metadata column (via \code{mcols}) to use}

\item{remove_na}{Take out NAs? Not done by default to preserve vector length}
}
\value{
output Vector of converted ids
}
\description{
Convert row names to metadata identifiers
}
