% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro.R
\name{dendro}
\alias{dendro}
\title{The server function of the dendrogram module}
\usage{
dendro(input, output, session, eselist)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\description{
This module will produce a sample clustering dendrogram based on 
user-selected parameters of row (e.g. gene) and column (sample) selection 
provided by the \code{selectmatrix} module, as well distance matrix 
generation and clustering method.
}
\details{
This function is not called directly, but rather via callModule() (see 
example).
}
\examples{
callModule(dendro, 'myid', eselist)
}
\keyword{shiny}
