% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro.R
\name{dendro}
\alias{dendro}
\title{The server function of the dendrogram module}
\usage{
dendro(input, output, session, ses)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{ses}{List of structuredExperiment objects with assay and experimental
data, with additional information in the metadata() slot}
}
\description{
This function is not called directly, but rather via callModule() (see 
example).
}
\examples{
callModule(dendro, 'dendro', ses)
}
\keyword{shiny}

