% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro.R
\name{dendroInput}
\alias{dendroInput}
\title{The input function of the dendrogram module}
\usage{
dendroInput(id, eselist)
}
\arguments{
\item{id}{Submodule namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\value{
output An HTML tag object that can be rendered as HTML using 
  as.character()
}
\description{
This module will produce a sample clustering dendrogram based on 
user-selected parameters of row (e.g. gene) and column (sample) selection 
provided by the \code{selectmatrix} module, as well distance matrix 
generation and clustering method.
}
\details{
This funcion provides the form elements to control the display
}
\examples{
library(shinyngs)
data(zhangneurons)
dendroInput('myid', zhangneurons)
}
\keyword{shiny}
