% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro.R
\name{dendroInput}
\alias{dendroInput}
\title{The input function of the dendrogram module}
\usage{
dendroInput(id, eses)
}
\arguments{
\item{id}{Submodule namespace}

\item{eses}{List of ExploratorySummarizedExperiment objects with assay and
experimental data}
}
\value{
output An HTML tag object that can be rendered as HTML using 
  as.character()
}
\description{
This provides the form elements to control the pca display
}
\examples{
dendroInput(ns('boxplot'), eses)
}
\keyword{shiny}

