% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro.R
\name{dendroOutput}
\alias{dendroOutput}
\title{The output function of the dendro module}
\usage{
dendroOutput(id)
}
\arguments{
\item{id}{Submodule namespace}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
This module will produce a sample clustering dendrogram based on 
user-selected parameters of row (e.g. gene) and column (sample) selection 
provided by the \code{selectmatrix} module, as well distance matrix 
generation and clustering method.
}
\details{
This provides actual dendrogram plot element for display by applications
}
\examples{
dendroOutput('myid')
}
\keyword{shiny}

