% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexseqtable.R
\name{dexseqtable}
\alias{dexseqtable}
\title{The server function of the dexseqtable module}
\usage{
dexseqtable(
  input,
  output,
  session,
  eselist,
  allow_filtering = TRUE,
  getDEUGeneID = NULL,
  show_controls = TRUE,
  page_length = 15,
  link_to_deu_plot = TRUE
)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}

\item{allow_filtering}{Allow user filtering of results (default: TRUE)?
Deactivated by the \code{dexseqplot} module, which uses it for showing
gene-specific results.}

\item{getDEUGeneID}{Reactive expression returning a gene ID.}

\item{show_controls}{Passed to \code{\link{simpletable}}, spcifies whether 
the various \code{datatables} controls are displayed (default: TRUE).}

\item{page_length}{Passed to \code{\link{simpletable}}, spcifies the number 
of rows to display (default: 15).}

\item{link_to_deu_plot}{Link label fields to the plots produced by 
\code{\link{dexseqplot}}? (default: TRUE)}
}
\description{
This module produces a differential exon usage table based on the output
\code{DEXSeqResults} object of the DEXSeq package.
}
\details{
For the table to be displayed, the \code{dexseq_results} slot must be filled
on at least one of the component \code{ExploratorySummarizedExperiment} objects
of the input \code{ExploratorySummarizedExperimentList}. 

\code{dexseq_results} must be a list of \code{DEXSeqResults} objects corresponding
to the contrasts listed in the \code{contrasts} slot of the 
\code{ExploratorySummarizedExperiment}.

This function is not called directly, but rather via callModule() (see 
example).
}
\examples{
callModule(dexseqtable, 'dexseqtable', eselist)
}
\keyword{shiny}
