% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialsummary.R
\name{differentialsummaryOutput}
\alias{differentialsummaryOutput}
\title{Output function for the differential summary module}
\usage{
differentialsummaryOutput(id)
}
\arguments{
\item{id}{Module namespace.}
}
\value{
Differential count plot and its backing table.
}
\description{
Output function for the differential summary module
}
\keyword{shiny}
