% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialtable.R
\name{differentialtable}
\alias{differentialtable}
\title{The server function of the differentialtable module}
\usage{
differentialtable(input, output, session, eselist)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\description{
This module provides information on the comparison betwen pairs of groups 
defined in a 'contrasts' slot of a ExploratorySummarizedExperimentList
}
\details{
This function is not called directly, but rather via callModule() (see 
example).

Essentially this funnction uses the \code{contrasts} module to group samples
and calculate fold changes, adding test statistics where available.
}
\examples{
callModule(differentialtable, 'differentialtable', eselist)
}
\keyword{shiny}
