% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialtable.R
\name{differentialtableInput}
\alias{differentialtableInput}
\title{The UI input function of the differentialtable module}
\usage{
differentialtableInput(id, eses)
}
\arguments{
\item{id}{Submodule namespace}

\item{eses}{List of ExploratorySummarizedExperiment objects with assay and
experimental data}
}
\value{
output An HTML tag object that can be rendered as HTML using 
  as.character()
}
\description{
This module provides information on the comparison betwen pairs of groups 
defined in a 'tests' slot of the ExploratorySummarizedExperiment
}
\details{
Leverages the \code{simpletable} module
}
\examples{
differentialtableInput('experiment', eses)
}
\keyword{shiny}

