% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{eselistfromConfig}
\alias{eselistfromConfig}
\title{Build an ExploratorySummarisedExperimentList from a description provided in a list}
\usage{
eselistfromConfig(config, log2_assays, log2_threshold = 30)
}
\arguments{
\item{config}{Hierachical named list with input components. See \code{eselistFromYAML} for detail.}

\item{log2_assays}{A string parameter that can be NULL, empty, or a non-empty string.
If NULL: log2 transformation will be guessed based on input assays.
If empty: no log2 transformation will be applied.
If non-empty: log2 transformation will be applied unconditionally to specified assays.}

\item{log2_threshold}{A numeric threshold to determine if the matrix should be log-transformed.
This is only checked if should_transform is NULL.}
}
\value{
out An ExploratorySummarizedExperimentList object suitable for passing to \code{\link{prepareApp}}
}
\description{
Build an ExploratorySummarisedExperimentList from a description provided in a list
}
