% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{eselistfromConfig}
\alias{eselistfromConfig}
\title{Build an ExploratorySummarisedExperimentList from a description provided in a list}
\usage{
eselistfromConfig(config)
}
\arguments{
\item{config}{Hierachical named list with input components. See \code{eselistFromYAML} for detail.}
}
\value{
out An ExploratorySummarizedExperimentList object suitable for passing to \code{\link{prepareApp}}
}
\description{
Build an ExploratorySummarisedExperimentList from a description provided in a list
}
