% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimenttable.R
\name{experimenttable}
\alias{experimenttable}
\title{The server function of the experimenttable module}
\usage{
experimenttable(id, eselist)
}
\arguments{
\item{id}{Module namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\description{
This function is called directly, using the same id as its UI counterpart,
and wraps its logic in \code{moduleServer()} (see example). Essentially this just passes the results of \code{colData()}
applied to the specified SummarizedExperiment object to the
\code{simpletable} module
}
\examples{
experimenttable("experimenttable", eselist)

# Almost certainly used via application creation

data(zhangneurons)
app <- prepareApp("experimenttable", zhangneurons)
shiny::shinyApp(ui = app$ui, server = app$server)

}
\keyword{shiny}
