% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimenttable.R
\name{experimenttable}
\alias{experimenttable}
\title{The server function of the experimenttable module}
\usage{
experimenttable(input, output, session, eselist)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\description{
This function is not called directly, but rather via callModule() (see 
example). Essentially this just passes the results of \code{colData()} 
applied to the specified SummarizedExperiment object to the 
\code{simpletable} module
}
\examples{
callModule(experimenttable, 'experimenttable', eselist)

# Almost certainly used via application creation

data(zhangneurons)
app <- prepareApp('experimenttable', zhangneurons)
shiny::shinyApp(ui = app$ui, server = app$server)
}
\keyword{shiny}

