% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimenttable.R
\name{experimenttableInput}
\alias{experimenttableInput}
\title{The UI input function of the experimenttable module
 
This module produces a simple table of the \code{colData()} in a
SummarizedExperiment object. If more than one of these objects were 
specified, a select box should appear to pick the desired one for display.}
\usage{
experimenttableInput(id, ses)
}
\arguments{
\item{id}{Submodule namespace}

\item{ses}{List of structuredExperiment objects with assay and experimental
data, with additional information in the metadata() slot}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
Leverages the \code{simpletable} module
}
\examples{
experimentableInput('experiment', ses)
}
\keyword{shiny}

