% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimenttable.R
\name{experimenttableInput}
\alias{experimenttableInput}
\title{The UI input function of the experimenttable module}
\usage{
experimenttableInput(id, eselist)
}
\arguments{
\item{id}{Submodule namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\value{
output An HTML tag object that can be rendered as HTML using 
  as.character()
}
\description{
This module produces a simple table of the \code{colData()} in a 
SummarizedExperiment object. If more than one of these objects were 
specified, a select box should appear to pick the desired one for display.
}
\details{
Leverages the \code{simpletable} module
}
\examples{
experimentableInput('experiment', eselist)

# Almost certainly used via application creation

data(zhangneurons)
app <- prepareApp('experimenttable', zhangneurons)
shiny::shinyApp(ui = app$ui, server = app$server)
}
\keyword{shiny}

