% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploratorySummarizedExperiment-class.R
\docType{class}
\name{ExploratorySummarizedExperiment-class}
\alias{ExploratorySummarizedExperiment-class}
\title{The ExploratorySummarizedExperiment class}
\description{
The ExploratorySummarizedExperiment class
}
\section{Slots}{

\describe{
\item{\code{idfield}}{character.}

\item{\code{entrezgenefield}}{character.}

\item{\code{labelfield}}{character.}

\item{\code{tests}}{}
}}

