% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploratorySummarizedExperiment-class.R
\docType{class}
\name{ExploratorySummarizedExperiment-class}
\alias{ExploratorySummarizedExperiment-class}
\title{The ExploratorySummarizedExperiment class}
\description{
Subclass of SummarizedExperiment if present in the SummarizedExperiment
package (newer versions of Bioconductor have moved this from GenomicRanges), 
otherwise of SummarizedExperiment0.
}
\section{Slots}{

\describe{
\item{\code{idfield}}{character.}

\item{\code{entrezgenefield}}{character.}

\item{\code{labelfield}}{character.}

\item{\code{contrast_stats}}{list.}

\item{\code{read_reports}}{list.}
}}

