% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploratorySummarizedExperiment-class.R
\name{ExploratorySummarizedExperiment}
\alias{ExploratorySummarizedExperiment}
\title{Title}
\usage{
ExploratorySummarizedExperiment(assays, colData, annotation, idfield,
  labelfield = character(), entrezgenefield = character(), tests = list())
}
\arguments{
\item{assays}{}

\item{colData}{}

\item{annotation}{}

\item{idfield}{}

\item{labelfield}{}

\item{entrezgenefield}{}

\item{tests}{}

\item{group_vars}{}

\item{default_groupvar}{}

\item{contrasts}{}

\item{url_roots}{}

\item{geneset_files}{}
}
\value{
output An ExploratoryRangedSummarizedExperient object
}
\description{
Title
}

