% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploratorySummarizedExperimentList-class.R
\docType{class}
\name{ExploratorySummarizedExperimentList-class}
\alias{ExploratorySummarizedExperimentList-class}
\title{The ExploratorySummaizedExperimentList class}
\description{
The ExploratorySummaizedExperimentList class
}
\section{Slots}{

\describe{
\item{\code{title}}{character.}

\item{\code{author}}{character.}

\item{\code{description}}{character.}

\item{\code{group_vars}}{character.}

\item{\code{default_groupvar}}{character.}

\item{\code{contrasts}}{list.}

\item{\code{url_roots}}{list.}

\item{\code{gene_sets}}{list.}

\item{\code{read_reports}}{list.}

\item{\code{ensembl_species}}{character.}
}}

