% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploratorySummarizedExperimentList-class.R
\docType{class}
\name{ExploratorySummarizedExperimentList-class}
\alias{ExploratorySummarizedExperimentList-class}
\alias{[,ExploratorySummarizedExperimentList,ANY,missing,ANY-method}
\alias{[,ExploratorySummarizedExperimentList,numeric,missing,ANY-method}
\alias{[,ExploratorySummarizedExperimentList,logical,missing,ANY-method}
\title{The ExploratorySummaizedExperimentList class}
\usage{
\S4method{[}{ExploratorySummarizedExperimentList,ANY,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{ExploratorySummarizedExperimentList,numeric,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{ExploratorySummarizedExperimentList,logical,missing,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{\code{ExploratorySummarizedExperimentList} object}

\item{i}{boolean index for the ExploratorySummarizedExperimentList list}

\item{j}{not used}

\item{...}{additional arguments not used here}

\item{drop}{not used}
}
\description{
The ExploratorySummaizedExperimentList class

Extract parts of ExploratorySummarizedExperimentList.

Extract parts of ExploratorySummarizedExperimentList.

Extract parts of ExploratorySummarizedExperimentList.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{list.}

\item{\code{title}}{character.}

\item{\code{author}}{character.}

\item{\code{description}}{character.}

\item{\code{static_pdf}}{character.}

\item{\code{group_vars}}{character.}

\item{\code{default_groupvar}}{character.}

\item{\code{contrasts}}{list.}

\item{\code{url_roots}}{list.}

\item{\code{gene_sets}}{list.}

\item{\code{gene_set_id_type}}{character.}

\item{\code{ensembl_species}}{character.}
}}

