% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploratorySummarizedExperimentList-class.R
\name{ExploratorySummarizedExperimentList}
\alias{ExploratorySummarizedExperimentList}
\title{Title}
\usage{
ExploratorySummarizedExperimentList(eses, title = "", author = "",
  description = "", group_vars = character(),
  default_groupvar = character(), contrasts = list(), url_roots = list(),
  geneset_files = list())
}
\arguments{
\item{eses}{}

\item{title}{}

\item{author}{}

\item{description}{}

\item{group_vars}{}

\item{default_groupvar}{}

\item{contrasts}{}

\item{url_roots}{}

\item{geneset_files}{}
}
\value{
output An ExploratorySummarizedExperimentList
}
\description{
Title
}

