% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{fieldSets}
\alias{fieldSets}
\title{Create sets of fields for display}
\usage{
fieldSets(id, fieldset_list, open = NULL, use_shinybs = TRUE)
}
\arguments{
\item{id}{ID field to apply to the overall container}

\item{fieldset_list}{A named list, each element containing one or more 
fields.}

\item{open}{Only applicable for output with shinyBS, controls which panels
are open by default. In most cases all should be left open (the default),
since shiny doesn't receive the inputs of fields in collapsed elements.}

\item{use_shinybs}{Use collapsible panels from shinyBS if installed}
}
\value{
list
}
\description{
Shiny apps can get cluttered with many inputs. This method wraps sets of
fields in either a \code{bsCollapse} from \code{shinyBS} (if installed) or
a simple div element with a title and class 'shinyngsFieldset' (which can 
then be used for styling)
}
