% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{foldChange}
\alias{foldChange}
\title{Fold change between two vectors}
\usage{
foldChange(vec1, vec2)
}
\arguments{
\item{vec1}{First vector}

\item{vec2}{Second vector}
}
\value{
Vector of fold changes
}
\description{
Fold change between two vectors
}

