% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldchangeplot.R
\name{foldchangeplot}
\alias{foldchangeplot}
\title{The server function of the \code{foldchangeplot} module}
\usage{
foldchangeplot(input, output, session, eselist)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\description{
This module is for making scatter plots comparing pairs of groups defined in
a 'tests' slot of the ExploratorySummarizedExperiment
}
\details{
This function is not called directly, but rather via callModule() (see 
example).
}
\examples{
callModule(foldchangeplot, 'foldchangeplot', eselist)
}
\keyword{shiny}

