% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldchangeplot.R
\name{foldchangeplotInput}
\alias{foldchangeplotInput}
\title{The UI input function of the \code{foldchangeplot} module}
\usage{
foldchangeplotInput(id, eselist)
}
\arguments{
\item{id}{Submodule namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\value{
output An HTML tag object that can be rendered as HTML using 
  as.character()
}
\description{
This module is for making scatter plots comparing pairs of groups defined in
a 'contrasts' slot of the ExploratorySummarizedExperimentList
}
\details{
Leverages the \code{scatterplot} module
}
\examples{
data(zhangneurons)
foldchangeplotInput('myid', zhangneurons)

# Almost certainly used via application creation

data(zhangneurons)
app <- prepareApp('foldchangeplot', zhangneurons)
shiny::shinyApp(ui = app$ui, server = app$server)
}
\keyword{shiny}
