% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldchangeplot.R
\name{foldchangeplotOutput}
\alias{foldchangeplotOutput}
\title{The output function of the \code{foldchangeplot} module}
\usage{
foldchangeplotOutput(id)
}
\arguments{
\item{id}{Module namespace}
}
\value{
output An HTML tag object that can be rendered as HTML using
as.character()
}
\description{
This module is for making scatter plots comparing pairs of groups defined in
a 'contrasts' slot of the ExploratorySummarizedExperimentList
}
\details{
Leverages the \code{scatterplot} module
}
\examples{
foldchangeplotOutput("myid")

data(zhangneurons)
app <- prepareApp("foldchangeplot", zhangneurons)
shiny::shinyApp(ui = app$ui, server = app$server)

}
\keyword{shiny}
