% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{geneInput}
\alias{geneInput}
\title{The input function of the gene module}
\usage{
geneInput(id, ses)
}
\arguments{
\item{id}{Submodule namespace}

\item{ses}{List of structuredExperiment objects with assay and experimental
data, with additional information in the metadata() slot}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
The gene module picks specified rows out the assay data, either simply by 
id or label. This is used to create a gene-centric info page.
}
\details{
Inputs are a gene label and a variable to color by (where available)
}
\examples{
geneInput(ns('gene'), ses)
}
\keyword{shiny}

