% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{geneOutput}
\alias{geneOutput}
\title{The input function of the gene module}
\usage{
geneOutput(id, eselist)
}
\arguments{
\item{id}{Submodule namespace}

\item{eselist}{List of structuredExperiment objects with assay and experimental 
data}
}
\value{
output An HTML tag object that can be rendered as HTML using 
  as.character()
}
\description{
The gene module picks specified rows out the assay data, either simply by id
or label. This is used to create a gene-centric info page.
}
\details{
Outputs are a bar plot and a table contrast data for this gene
}
\examples{
geneOutput('gene', eselist)
}
\keyword{shiny}
