% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneselect.R
\name{geneselect}
\alias{geneselect}
\title{The server function of the geneselect module}
\usage{
geneselect(input, output, session, se, transcriptfield, entrezgenefield,
  genefield, geneset_files = NULL, selectSamples)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{transcriptfield}{The main identifier for the rows in the assay data.
This could be transcript ID, but also probe etc.}

\item{entrezgenefield}{The column of annotation containing Entrez gene IDs}

\item{genefield}{The gene ID type in annotation by which results are keyed}

\item{geneset_files}{(optional) A named list of .gmt gene set files as might be 
derived from MSigDB}

\item{getMatrix}{A reactive expression that provides a matrix of values.
This will be used to calculate variance, for example, where that is used 
as a row-filtering criterion.}
}
\value{
output A list of reactive functions for interrogating the selected
rows.
}
\description{
This module provides controls for selecting genes (matrix rows) by various
criteria such as variance and gene set.
}
\examples{
geneselect_functions <- callModule(geneselect, 'heatmap', se, transcriptfield, entrezgenefield, genefield, geneset_files, getMatrix=selectColumns)
}
\keyword{shiny}

