% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneselect.R
\name{geneselect}
\alias{geneselect}
\title{The server function of the geneselect module}
\usage{
geneselect(input, output, session, getExperiment, var_n = 50, var_max = 500,
  selectSamples, assay)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{getExperiment}{Reactive expression which returns a
StructuredExperiment object with assay and experimental data, with
additional information in the metadata() slot}

\item{var_n}{The number of rows to select when doing so by variance. Default = 50}

\item{var_max}{The maximum umber of rows to select when doing so by variance. 
Default = 500}

\item{selectSamples}{A reactive expression that provides a vector of samples
to use, e.g. in row-wise variance calculation}
}
\value{
output A list of reactive functions for interrogating the selected
rows.
}
\description{
This module provides controls for selecting genes (matrix rows) by various
criteria such as variance and gene set.
}
\examples{
geneselect_functions <- callModule(geneselect, 'heatmap', getExperiment, getMatrix=selectColumns)
}
\keyword{shiny}

