% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneselect.R
\name{geneselect}
\alias{geneselect}
\title{The server function of the geneselect module}
\usage{
geneselect(
  input,
  output,
  session,
  eselist,
  getExperiment,
  var_n = 50,
  var_max = 500,
  selectSamples = reactive({
     colnames(getExperiment())
 }),
  getAssay,
  provide_all = TRUE,
  provide_none = FALSE,
  default = NULL
)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}

\item{getExperiment}{Reactive expression which returns a
ExploratorySummarizedExperiment object with assay and experimental data}

\item{var_n}{The number of rows to select when doing so by variance. Default
= 50}

\item{var_max}{The maximum umber of rows to select when doing so by variance.
Default = 500}

\item{selectSamples}{A reactive expression that provides a vector of samples
to use, e.g. in row-wise variance calculation}

\item{getAssay}{A reactive expression providing the current assay selection.}

\item{provide_all}{Allow the 'all rows' selection in the UI? Means we don't
have to calculate variance so the display is quicker, but it's a bad idea
for e.g. heatmaps where the visual scales by the number of rows.}

\item{provide_none}{Allow the 'none' selection in the UI to make row
selection optional.}

\item{default}{Default gene selection method}
}
\value{
output A list of reactive functions for interrogating the selected
  rows.
}
\description{
This module provides controls for selecting genes (matrix rows) by various
criteria such as variance and gene set.
}
\examples{
geneselect_functions <- callModule(geneselect, "heatmap", getExperiments)

}
\keyword{shiny}
