% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesetbarcodeplot.R
\name{genesetbarcodeplotOutput}
\alias{genesetbarcodeplotOutput}
\title{The output function of the genesetbarcodeplot module}
\usage{
genesetbarcodeplotOutput(id)
}
\arguments{
\item{id}{Module namespace}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
This module leverages gene sets stored in the \code{gene_sets} slot of an
\code{ExploratorySummarizedExperimentList} object to produce barcode plots
using Limma's \code{\link[limma]{barcodeplot}} function. Genes are ranked by
fold changes calculated with the \code{contrasts} module, and FDR values 
from the \code{gene_set_analyses} slot of the selected 
\code{ExploratorySummarizedExperiment} are displayed where provided.
}
\examples{
genesetbarcodeplotOutput('experiment')

# Almost certainly used via application creation

app <- prepareApp('genesetbarcodeplot', zhangneurons)
shiny::shinyApp(ui = app$ui, server = app$server)
}
\keyword{shiny}

