% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesetselect.R
\name{genesetselectInput}
\alias{genesetselectInput}
\title{The UI function of the genesetselect module}
\usage{
genesetselectInput(id, multiple = TRUE)
}
\arguments{
\item{id}{Submodule namespace}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
The gene set module is for adding a gene set filter to displays. A
\code{\link[shiny]{selectizeInput}} is used for performance reasons, 
providing an autocomplete field for selecting from a list that could stretch 
to thousands of entries. This would be difficult to do client-side using a 
standard select field.
}
\examples{
genesetselectInput('myid')
}
\keyword{shiny}
