% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{ggplot_boxplot}
\alias{ggplot_boxplot}
\title{Make a boxplot with coloring by experimental variable}
\usage{
ggplot_boxplot(plotmatrix, experiment, colorby = NULL, palette = NULL,
  expressiontype = "expression", whisker_distance = 1.5)
}
\arguments{
\item{plotmatrix}{Expression/ other data matrix}

\item{experiment}{Annotation for the columns of plotmatrix}

\item{colorby}{Column name in \code{experiment} specifying how boxes should be colored}

\item{palette}{Palette of colors, one for each unique value derived from 
\code{colorby}.}

\item{expressiontype}{Expression type for use in y axis label}

\item{whisker_distance}{Passed to \code{\link[ggplot2]{geom_boxplot}} as 
\code{coef}, controlling the length of the whiskers. See documentation of 
that function for more info (default: 1.5).}
}
\value{
output A \code{ggplot} output
}
\description{
A simple function using \code{ggplot2} to make a sample boxplot
}
\examples{
data(airway, package = 'airway')
ggplot_boxplot(assays(airway)[[1]], data.frame(colData(airway)), colorby = 'dex')
}
\keyword{keywords}
