% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{ggplot_boxplot}
\alias{ggplot_boxplot}
\title{Make a boxplot with coloring by experimental variable}
\usage{
ggplot_boxplot(plotmatrix, experiment, colorby = NULL,
  expressiontype = "Normalised counts per million", whisker_distance = 1.5)
}
\arguments{
\item{plotmatrix}{Expression/ other data matrix}

\item{experiment}{Annotation for the columns of plotmatrix}

\item{colorby}{Column name in \code{experiment} specifying how boxes should be colored}

\item{expressiontype}{Expression type for use in y axis label}
}
\value{
output A \code{ggplot} output
}
\description{
A simple function using \code{ggplot2} to make a sample boxplot
}
\examples{
ggplot_boxplot(selectMatrix(), selectColData(), colorBy())
}
\keyword{keywords}

