% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{ggplot_densityplot}
\alias{ggplot_densityplot}
\title{Make a static density plot with ggplot2}
\usage{
ggplot_densityplot(
  plotmatrices,
  experiment,
  colorby = NULL,
  palette = NULL,
  expressiontype = "expression",
  base_size = 16,
  palette_name = "Set1"
)
}
\arguments{
\item{plotmatrices}{Expression/ other data matrix, or named list thereof}

\item{experiment}{Annotation for the columns of plotmatrix}

\item{colorby}{Column name in \code{experiment} specifying how boxes should be colored}

\item{palette}{Palette of colors, one for each unique value derived from
\code{colorby}.}

\item{expressiontype}{Expression type for use in y axis label}

\item{base_size}{Passed to ggplot's \code{theme()}}

\item{palette_name}{Valid R color palette name}
}
\value{
output A \code{ggplot} output
}
\description{
Make a static density plot with ggplot2
}
