% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{ggplotify}
\alias{ggplotify}
\title{Reshape data to the way \code{ggplot2} likes it}
\usage{
ggplotify(plotmatrices, experiment, colorby = NULL)
}
\arguments{
\item{plotmatrices}{A matrix of values, e.g. expression data}

\item{experiment}{A data frame with rows matching the columns of
\code{matrix}}

\item{colorby}{An optional string specifying a column from \code{experiment}
that will be used to set a color column in the reshaped output.}
}
\value{
A reshaped data frame
}
\description{
Reshape data to the way \code{ggplot2} likes it
}
\examples{
plotdata <- ggplotify(as.matrix(plotmatrix), experiment, colorby)

}
