% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupby.R
\name{groupby}
\alias{groupby}
\title{The server function of the groupby module}
\usage{
groupby(input, output, session, getExperiment, group_label = "Group by",
  multiple = FALSE)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{getExperiment}{A reactive that returns a SummarizedExperiment}

\item{group_label}{A label for the grouping field}

\item{multiple}{Produces a checkbox group if true, a select box if false}
}
\value{
output A list of two reactive functions: getPathwayNames() and 
getPathwayGenes() which will be used by other modules.
}
\description{
The groupby module provides a UI element to choose from the group_vars in
a SummarizedExperment. Useful for coloring in a PCA etc
}
\examples{
geneset_functions <- callModule(groupby, 'heatmap', getExperiment)
}
\keyword{shiny}

