% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap}
\alias{heatmap}
\title{The server function of the heatmap module}
\usage{
heatmap(input, output, session, se, transcriptfield, entrezgenefield, genefield,
  geneset_files = NULL)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{se}{StructuredExperiment object with assay and experimental data}

\item{transcriptfield}{The main identifier for the rows in the assay data.
This could be transcript ID, but also probe etc.}

\item{entrezgenefield}{The column of annotation containing Entrez gene IDs}

\item{genefield}{The gene ID type in annotation by which results are keyed}

\item{geneset_files}{A named list of .gmt gene set files as might be 
derived from MSigDB}
}
\description{
This function is not called directly, but rather via callModule() (see 
example).
}
\details{
This function assumes that the gene sets have one gene ID (e.g. Entrez)
which need to be converted to another (e.g. Symbol, Ensembl gene ID).
This would be common when dealign with MSigDB gene sets, for example.
}
\examples{
callModule(heatmap, 'heatmap', se, params$transcriptfield, params$entrezgenefield, params$genefield, geneset_files=params$geneset_files)
}
\keyword{shiny}

