% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap}
\alias{heatmap}
\title{The server function of the heatmap module}
\usage{
heatmap(input, output, session, ses)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{ses}{List of structuredExperiment objects with assay and experimental
data, with additional information in the metadata() slot}
}
\description{
This function is not called directly, but rather via callModule() (see 
example).
}
\details{
This function assumes that the gene sets have one gene ID (e.g. Entrez)
which need to be converted to another (e.g. Symbol, Ensembl gene ID).
This would be common when dealign with MSigDB gene sets, for example.
}
\examples{
callModule(heatmap, 'heatmap', se, params$transcriptfield, params$entrezgenefield, params$genefield, geneset_files=params$geneset_files)
}
\keyword{shiny}

