% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmapInput}
\alias{heatmapInput}
\title{The input function of the heatmap module}
\usage{
heatmapInput(id, ses)
}
\arguments{
\item{id}{Submodule namespace}

\item{ses}{List of structuredExperiment objects with assay and experimental
data, with additional information in the metadata() slot}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
This provides the form elements to control the heatmap display
}
\examples{
heatmapInput('heatmap', se, group_vars, default_groupvar
}
\keyword{shiny}

