% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmapInput}
\alias{heatmapInput}
\title{The input function of the heatmap module}
\usage{
heatmapInput(id, se, group_vars, default_groupvar)
}
\arguments{
\item{id}{Submodule namespace}

\item{se}{StructuredExperiment object with assay and experimental data}

\item{group_vars}{The variables from the structured experiment that should
be used to control sample grouping in the plot}

\item{The}{default grouping variable to use}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
This provides the form elements to control the heatmap display
}
\examples{
heatmapInput("heatmap", se, group_vars, default_groupvar
}
\keyword{shiny}

