% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmapInput}
\alias{heatmapInput}
\title{The input function of the heatmap module}
\usage{
heatmapInput(id, eses, type = "expression")
}
\arguments{
\item{id}{Submodule namespace}

\item{eses}{List of ExploratorySummarizedExperiment objects with assay and
experimental data}

\item{type}{The type of heatmap that will be made. 'expression', 'samples' or
'pca'}
}
\value{
output An HTML tag object that can be rendered as HTML using 
  as.character()
}
\description{
This provides the form elements to control the heatmap display
}
\examples{
heatmapInput('heatmap', ese, group_vars, default_groupvar)
}
\keyword{shiny}

