% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illuminaarray.R
\name{illuminaarray}
\alias{illuminaarray}
\title{The server function of the illuminaarray module}
\usage{
illuminaarray(id, eselist)
}
\arguments{
\item{id}{Module namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\description{
This function is called directly, using the same id as its UI counterpart,
and wraps its logic in \code{moduleServer()} (see example).
}
\examples{
illuminaarray("illuminaarray", eselist)

}
\keyword{shiny}
