% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illuminaarray.R
\name{illuminaarray}
\alias{illuminaarray}
\title{The server function of the illuminaarray module}
\usage{
illuminaarray(input, output, session, eselist)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\description{
This function is not called directly, but rather via callModule() (see 
example).
}
\examples{
callModule(illuminaarray, 'illuminaarray', eselist)
}
\keyword{shiny}
