% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illuminaarrayqc.R
\name{illuminaarrayqc}
\alias{illuminaarrayqc}
\title{The server function of the illuminaarrayqc module}
\usage{
illuminaarrayqc(input, output, session, eselist)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\description{
This module plots control probes from an illumina microarray experiment.
}
\details{
This function is not called directly, but rather via callModule() (see 
example).
}
\examples{
callModule(illuminaarayqc, 'myid', eselist)
}
\keyword{shiny}

