% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{interactiveHeatmap}
\alias{interactiveHeatmap}
\title{Make a ineractive heatmap with d3heatmap}
\usage{
interactiveHeatmap(plotmatrix, sample_annotation, group_vars,
  cluster_rows = TRUE, cluster_cols = FALSE, scale = "row", ...)
}
\arguments{
\item{plotmatrix}{Expression/ other data matrix}

\item{sample_annotation}{Annotation for the columns of plotmatrix}

\item{group_vars}{Grouping variables to be used from 
sample_annotation}

\item{cluster_rows}{Make row-wise dendrogram?}

\item{cluster_cols}{Make column-wise dendrogram?}

\item{scale}{row, column or none}
}
\value{
output A plot as produced by pheatmap()
}
\description{
This is a generic function which may be useful outside of this package. It
produces a heatmap based on an expression matrix and accompanying 
experiment data in the form of a frame, \code{d3heatmap()}.
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

