% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{interactiveHeatmap}
\alias{interactiveHeatmap}
\title{Make a ineractive heatmap with d3heatmap}
\usage{
interactiveHeatmap(plotmatrix, displaymatrix, sample_annotation,
  cluster_rows = TRUE, cluster_cols = FALSE, scale = "row", row_labels,
  colors = colorRampPalette(rev(RColorBrewer::brewer.pal(n = 7, name =
  "RdYlBu")))(100), cexCol = 0.7, cexRow = 0.7, ...)
}
\arguments{
\item{plotmatrix}{Expression/ other data matrix}

\item{displaymatrix}{A matrix of values that might be displayed in cells}

\item{cluster_rows}{Cluster rows?}

\item{cluster_cols}{Cluster columns?}

\item{scale}{'row', 'column' or none}

\item{row_labels}{Vector labels to use for rows}

\item{row_height}{The height to use for each row}
}
\value{
output A plot as produced by pheatmap()
}
\description{
This is a generic function which may be useful outside of this package. It
produces a heatmap based on an expression matrix and accompanying 
experiment data in the form of a frame, \code{d3heatmap()}.
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

