% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{interleaveColumns}
\alias{interleaveColumns}
\title{Interleave the columns of two matrices of equal dimensions}
\usage{
interleaveColumns(mat1, mat2)
}
\arguments{
\item{mat1}{First numeric matrix}

\item{mat2}{Second numeric matrix}
}
\value{
output Interleaved matrix
}
\description{
Interleave the columns of two matrices of equal dimensions
}
