% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectmatrix.R
\name{linkMatrix}
\alias{linkMatrix}
\title{Add links to a table}
\usage{
linkMatrix(matrix, ese)
}
\arguments{
\item{matrix}{The input table}

\item{ese}{An ExploratorySummarizedExperiment object}
}
\value{
output Table with links added
}
\description{
Root URLs must be present in the \code{url_roots} slot of \code{se}
}

