% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectmatrix.R
\name{linkMatrix}
\alias{linkMatrix}
\title{Add links to a table}
\usage{
linkMatrix(matrix, url_roots)
}
\arguments{
\item{matrix}{The input table}

\item{url_roots}{A list with URL roots, with names matching columns of 
\code{matrix}}
}
\value{
output Table with links added
}
\description{
Root URLs must be present in the \code{url_roots} slot of \code{se}
}

