% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{makeContrastControl}
\alias{makeContrastControl}
\title{Make a select field for picking one or more contrasts}
\usage{
makeContrastControl(id, contrasts, contrast_numbers, summaries,
  multiple = FALSE, show_controls = TRUE, select_all = FALSE)
}
\arguments{
\item{id}{An id to apply to form elements}

\item{contrasts}{A list of lists specifying contrasts.}

\item{contrast_numbers}{A named vector of indices corresponding to 
\code{contrasts}.}

\item{summaries}{Summaries derived from \code{\link{summarizematrix}}, will
be used to select valid contrasts.}

\item{multiple}{Allow multiple contrasts to be selected?}

\item{show_controls}{Show controls? Setting to false will cause them to be 
hidden.}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
Make a select field for picking one or more contrasts
}

