% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{makeContrastFilterSet}
\alias{makeContrastFilterSet}
\title{Make a complete set of filters for a contrast: the contrast itself, fold
change, and where applicable p- and q- values.}
\usage{
makeContrastFilterSet(ns, ese, assay, summaries, contrasts, contrast_numbers,
  multiple, show_controls, default_min_foldchange = default_min_foldchange,
  index = "", filter_rows = TRUE, select_all_contrasts = FALSE)
}
\arguments{
\item{ns}{A namespace function (created with \code{\link[shiny]{NS}} to be 
used in creating field IDs.}

\item{ese}{ExploratorySummarizedExperiment object}

\item{assay}{Assay in \code{ese}}

\item{summaries}{Summaries derived from \code{\link{summarizematrix}}, will
be used to select valid contrasts.}

\item{contrasts}{A list of lists specifying contrasts.}

\item{contrast_numbers}{A named vector of indices corresponding to 
\code{contrasts}.}

\item{multiple}{Allow multiple contrasts to be selected? Passed to 
\code{\link{makeContrastControl}}.}

\item{show_controls}{Show controls? Setting to false will cause them to be 
hidden.}

\item{default_min_foldchange}{Default value for the fold change field}

\item{index}{Index. Will be used to differentiate mutiple copies of the
field set.}

\item{filter_rows}{Use fold change and p value etc to filter values?}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
Make a complete set of filters for a contrast: the contrast itself, fold
change, and where applicable p- and q- values.
}

