% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{makeHeatmap}
\alias{makeHeatmap}
\title{Make the heatmap}
\usage{
makeHeatmap(input, exprmatrix, experiment, group_vars = NULL,
  interactive = FALSE, ...)
}
\arguments{
\item{input}{Input object}

\item{exprmatrix}{An expression matrix}

\item{experiment}{A data frame containing experimental variables}

\item{...}{Arguments that should be passed through \code{annotatedHeatmap()}
to \code{pheatmap()}}
}
\value{
output Heatmap plot
}
\description{
Just an adapter to annotatedHeatmap(), which is designed to be generically
useful
}
\examples{
makeHeatmap(input, getHeatmapExpression(), selectColData(), main = matrixTitle())
}

