% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maplot.R
\name{maplotOutput}
\alias{maplotOutput}
\title{The output function of the \code{maplot} module}
\usage{
maplotOutput(id)
}
\arguments{
\item{id}{Module namespace}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
This module is for making scatter plots comparing pairs of groups 
defined in a 'tests' slot of the ExploratorySummarizedExperiment
}
\details{
Leverages the \code{scatterplot} module
}
\examples{
differentialtableOutput('experiment')
}
\keyword{shiny}

