% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modalInput}
\alias{modalInput}
\title{The input function for the \code{modal} module}
\usage{
modalInput(id, label, class, icon = "info-circle")
}
\arguments{
\item{id}{Modal ID. Must match that passed to \code{inlinehelpOutput}}

\item{label}{A label to use for the link}

\item{class}{A class to apply to the link}
}
\description{
This module uses modals from \code{shinyBS} to create overlaid text for the
current panel which displays when a link is clicked. Input and output
functions are placed in the output function of the calling module (see
example).
}
\details{
This is handy, for example when adding help text.
}
\examples{
modalInput(ns('dendro'), 'help'),
}

