% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modalServer}
\alias{modalServer}
\title{The server function of the \code{modal} module}
\usage{
modalServer(id, title, content = NULL)
}
\arguments{
\item{id}{Modal ID. Must match that passed to \code{modalInput}}

\item{title}{Title to show on the help modal}

\item{content}{Content to include in the modal. When \code{NULL} (the
default), Markdown is loaded from \code{inst/inlinehelp/<id>.md}.}
}
\description{
This module uses Shiny's \code{modalDialog()} to create overlaid text for the
current panel which displays when the link produced by \code{modalInput()} is
clicked. It is called from the server function of the calling module, using
the same id passed to \code{modalInput} (see example).
}
\details{
This is handy, for example when adding help text.
}
\examples{
modalServer("dendro", "Sample clustering dendrogram")

}
