% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{The server function of the pca module}
\usage{
pca(input, output, session, ses)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{ses}{List of structuredExperiment objects with assay and experimental
data, with additional information in the metadata() slot}
}
\description{
This module calculates a PCA and formats components and loadings for
display. It uses a common set of controls, generated with the 
\code{scatterplotcontrols} module, to power scatter plots for both 
components and loadings produced by the \code{scatterplots} module.
}
\details{
This function is not called directly, but rather via callModule() (see 
example).

Matrix and UI selection elements provided by the selectmatrix module
}
\examples{
callModule(pca, 'pca', ses)
}
\keyword{shiny}

