% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{The server function of the pca module}
\usage{
pca(input, output, session, se, transcriptfield, entrezgenefield, genefield,
  geneset_files = NULL)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{se}{StructuredExperiment object with assay and experimental data}

\item{transcriptfield}{The main identifier for the rows in the assay data.
This could be transcript ID, but also probe etc.}

\item{entrezgenefield}{The column of annotation containing Entrez gene IDs}

\item{genefield}{The gene ID type in annotation by which results are keyed}

\item{geneset_files}{(optional) A named list of .gmt gene set files as might be 
derived from MSigDB}
}
\description{
This function is not called directly, but rather via callModule() (see 
example).
}
\details{
Matrix and UI selection elements provided by the selectmatrix module
}
\examples{
callModule(pca, 'pca', se, params$transcriptfield, params$entrezgenefield, params$genefield, geneset_files = params$geneset_files)
}
\keyword{shiny}

