% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pcaOutput}
\alias{pcaOutput}
\title{The output function of the pca module}
\usage{
pcaOutput(id)
}
\arguments{
\item{id}{Module namespace}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
This provides actual pca element for display by applications
}
\examples{
pcaOutput('pca')
}
\keyword{shiny}

