% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdownload.R
\name{plotdownload}
\alias{plotdownload}
\title{The server function of the gene set module}
\usage{
plotdownload(input, output, session, makePlot, filename = "plot.png",
  plotHeight, plotWidth)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{makePlot}{A reactive for generating the plot}

\item{filename}{A filename (default = 'plot.png')}

\item{plotHeight}{A number or reactive for calculating the height of the plot}

\item{plotWidth}{A number or reactive for calculating the width of the plot}
}
\description{
The plotdownload module provides export functionality for panels with plots.
This will generally not be called directly, but by other modules
}
\examples{
callModule(plotdownload, 'heatmap', makePlot = plotHeatmap, filename = 'heatmap.png', plotHeight = plotHeight, plotWidth = plotWidth)
}
\keyword{shiny}

