% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{plotly_boxplot}
\alias{plotly_boxplot}
\title{Make a boxplot with coloring by experimental variable}
\usage{
plotly_boxplot(matrix, experiment, colorby, expressiontype = "expression")
}
\arguments{
\item{experiment}{Annotation for the columns of plotmatrix}

\item{colorby}{Column name in \code{experiment} specifying how boxes should be colored}

\item{expressiontype}{Expression type for use in y axis label}

\item{plotmatrix}{Expression/ other data matrix}
}
\value{
output A \code{plotly} output
}
\description{
A simple function using \code{plotly} to make a sample boxplot.
NOT CURRENTLY USED DUE TO RESOURCE REQUIREMENTS ON LARGE MATRICES
}
\keyword{keywords}
