% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{plotly_boxplot}
\alias{plotly_boxplot}
\title{Make a boxplot with coloring by experimental variable}
\usage{
plotly_boxplot(
  plotmatrices,
  experiment,
  colorby,
  palette = NULL,
  expressiontype = "expression",
  palette_name = "Set1"
)
}
\arguments{
\item{plotmatrices}{Expression/ other data matrix, or named list thereof}

\item{experiment}{Annotation for the columns of plotmatrix}

\item{colorby}{Column name in \code{experiment} specifying how boxes should be colored}

\item{palette}{Palette of colors, one for each unique value derived from
\code{colorby}.}

\item{expressiontype}{Expression type for use in y axis label}

\item{palette_name}{Valid R color palette name}
}
\value{
output A \code{plotly} output
}
\description{
A simple function using \code{plotly} to make a sample boxplot.
NOT CURRENTLY USED DUE TO RESOURCE REQUIREMENTS ON LARGE MATRICES
}
\keyword{keywords}
