% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{prepareApp}
\alias{prepareApp}
\title{Make UI and server functions for apps based on supplied data}
\usage{
prepareApp(type, se, params = list())
}
\arguments{
\item{type}{A string specifying the type of shiny app required (options: heatmap)}

\item{se}{A SummarizedExperiment object containing assay data (expression, 
counts...), sample data and annotation data for the rows.}

\item{params}{A list containing data and display options for the Shiny app}
}
\value{
output A list of length 2 containing: the UI and server components
}
\description{
Draws on various components (heatmaps, tables etc) to produce the UI and server
components of a variety of shiny apps, based on the type and data specified.
}
\examples{
app <- prepareApp("heatmap", se, params)
shinyApp(app$ui, app$server)
}
\keyword{shiny}

