% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{prettifyVariablename}
\alias{prettifyVariablename}
\title{Make machine variable names pretty for display}
\usage{
prettifyVariablename(vn)
}
\arguments{
\item{vn}{A string to be prettified}
}
\value{
output Prettified string
}
\description{
Just split out words using '_' and capitalise first letter
}
\examples{
vn <- 'ugly_name_of_thing'
> prettyifyVariablename(vn)
[1] "Ugly name of thing"
}

