% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{pushToList}
\alias{pushToList}
\title{Simple list push}
\usage{
pushToList(input_list, element)
}
\arguments{
\item{element}{The element to push}

\item{list}{A list to push to}
}
\value{
list with element pushed
}
\description{
Simple list push
}
\examples{
mylist <- pushToList(mylist, 'new element')
}
