% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{read_contrasts}
\alias{read_contrasts}
\title{Read and validate a contrasts file against sample metadata}
\usage{
read_contrasts(
  filename,
  samples,
  variable_column = "variable",
  reference_column = "reference",
  target_column = "target",
  blocking_column = "blocking"
)
}
\arguments{
\item{filename}{Contrasts file}

\item{samples}{Data frame of sample information}

\item{variable_column}{Column in contrasts file referencing sample sheet
column}

\item{reference_column}{Column in contrast file referencing reference level
of sample sheet variable}

\item{target_column}{Column in contrast file referencing target level of
sample sheet variable}

\item{blocking_column}{Colon-separated column in contrast file referencing
sample sheet variables to be used as blocking factors}
}
\value{
output Validated contrasts data frame
}
\description{
Read and validate a contrasts file against sample metadata
}
