% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{read_differential}
\alias{read_differential}
\title{Read tables of differential statistics}
\usage{
read_differential(
  filename,
  feature_id_column = NULL,
  pval_column = NULL,
  qval_column = NULL,
  fc_column = NULL,
  unlog_foldchanges = FALSE
)
}
\arguments{
\item{filename}{File name of file with table of differential statistics}

\item{feature_id_column}{Column of stats file with feature identifiers}

\item{pval_column}{Column of stats file with p values}

\item{qval_column}{Column of stats file with adjust p values/ q values}

\item{fc_column}{Column of stats with fold changes}
}
\value{
output Validated selected columns of differential stats files as a
  data frame
}
\description{
Read tables of differential statistics
}
