% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{read_matrix}
\alias{read_matrix}
\title{Read an expression matrix file and match to specified samples and features}
\usage{
read_matrix(
  matrix_file,
  sample_metadata,
  feature_metadata = NULL,
  sep = NULL,
  row.names = 1,
  guess_unlog = FALSE
)
}
\arguments{
\item{matrix_file}{Matrix file}

\item{sample_metadata}{Data frame of sample metadata}

\item{feature_metadata}{Data fraome of feature metadata}

\item{sep}{Sepaarator in matrix file}

\item{row.names}{Matrix column number or name containing feature identifiers}

\item{guess_unlog}{Should we guess the log status of matrices and unlog where
things seem logged?}
}
\value{
output Numeric matrix
}
\description{
Read an expression matrix file and match to specified samples and features
}
