% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readreports.R
\name{readreports}
\alias{readreports}
\title{Server function of the \code{readreports} module}
\usage{
readreports(input, output, session, eselist)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{eselist}{ExploratorySummarizedExperiment with \code{read_distribution}
slot filled}
}
\description{
Display plots and tables relating to read mapping, attrition during 
analysis etc.
}
