% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{rnaseq}
\alias{rnaseq}
\title{The server function of the rnaseq module}
\usage{
rnaseq(input, output, session, eselist)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\description{
This function is not called directly, but rather via callModule() (see 
example).
}
\examples{
callModule(rnaseq, 'rnaseq', eselist)
}
\keyword{shiny}
