% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{rnaseqInput}
\alias{rnaseqInput}
\title{The input function of the rnaseq module}
\usage{
rnaseqInput(id, ses, title = "")
}
\arguments{
\item{id}{Submodule namespace}

\item{ses}{List of structuredExperiment objects with assay and experimental
data, with additional information in the metadata() slot}

\item{title}{An optional title to display on the app}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
This provides the form elements to control the RNA-seq display
}
\details{
The rnaseq module is a combination of output from many modules (pca, boxplot
etc) to form a comprehensive analysis application.
}
\examples{
rnaseqInput('rnaseq', ses)
}
\keyword{shiny}

