% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{rnaseqInput}
\alias{rnaseqInput}
\title{The input function of the rnaseq module}
\usage{
rnaseqInput(id, eselist)
}
\arguments{
\item{id}{Submodule namespace}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\value{
output An HTML tag object that can be rendered as HTML using 
  as.character()
}
\description{
This provides the form elements to control the RNA-seq display
}
\details{
The rnaseq module is a combination of output from many modules (pca, boxplot 
etc) to form a comprehensive analysis application.
}
\examples{
rnaseqInput('rnaseq', eselist)
}
\keyword{shiny}
