% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowmetatable.R
\name{rowmetatable}
\alias{rowmetatable}
\title{The server function of the rowmetatable module}
\usage{
rowmetatable(input, output, session, eselist)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}
}
\description{
This function is not called directly, but rather via callModule() (see
example). Essentially this just passes the results of \code{colData()}
applied to the specified SummarizedExperiment object to the
\code{simpletable} module
}
\examples{
callModule(rowmetatable, "rowmetatable", eselist)

}
\keyword{shiny}
