% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{runPCA}
\alias{runPCA}
\title{Run a simple PCA analysis}
\usage{
runPCA(matrix)
}
\arguments{
\item{matrix}{Matrix (not logged)}
}
\value{
pca Output of the prcomp function
}
\description{
Common function for PCA-using parts of the app
}
\examples{
runPCA(mymatrix)
}
\keyword{shiny}
