% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleselect.R
\name{sampleselect}
\alias{sampleselect}
\title{The server function of the sampleselect module}
\usage{
sampleselect(
  input,
  output,
  session,
  eselist,
  getExperiment,
  allow_summarise = TRUE
)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}

\item{getExperiment}{Reactive expression that returns a
\code{ExploratorySummarizedExperiment} with assays and metadata. Usually a
result of a user selection}

\item{allow_summarise}{Boolean, show controls for matrix summarisation?}
}
\value{
output A list of reactive functions for interrogating the selected
samples/ columns.
}
\description{
This module provides controls for selecting matrix columns by sample or
group name.
}
\details{
This function is not called directly, but rather via callModule() (see
example).
}
\examples{
selectSamples <- callModule(sampleselect, "selectmatrix", getExperiment)

}
\keyword{shiny}
