% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleselect.R
\name{sampleselect}
\alias{sampleselect}
\title{The server function of the sampleselect module
 
This module provides controls for selecting matrix columns by sample or 
group name.}
\usage{
sampleselect(input, output, session, se)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{se}{StructuredExperiment object with assay and experimental data}
}
\value{
output A list of reactive functions for interrogating the selected
samples/ columns.
}
\description{
This function is not called directly, but rather via callModule() (see 
example).
}
\examples{
selectSamples <- callModule(sampleselect, "selectmatrix", se)
}
\keyword{shiny}

