% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleselect.R
\name{sampleselect}
\alias{sampleselect}
\title{The server function of the sampleselect module
 
This module provides controls for selecting matrix columns by sample or 
group name.}
\usage{
sampleselect(input, output, session, eselist, getExperiment)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}

\item{getExperiment}{Reactive expression that returns a
\code{ExploratorySummarizedExperiment} with assays and metadata. Usually a 
result of a user selection}
}
\value{
output A list of reactive functions for interrogating the selected
samples/ columns.
}
\description{
This function is not called directly, but rather via callModule() (see 
example).
}
\examples{
selectSamples <- callModule(sampleselect, 'selectmatrix', getExperiment)
}
\keyword{shiny}

