% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleselect.R
\name{sampleselectInput}
\alias{sampleselectInput}
\title{The UI input function of the sampleselect module}
\usage{
sampleselectInput(id, ese, select_samples = TRUE)
}
\arguments{
\item{id}{Submodule namespace}

\item{ese}{ExploratorySummarizedExperiment object with assay and experimental
data}
}
\value{
output An HTML tag object that can be rendered as HTML using 
  as.character()
}
\description{
This module provides controls for selecting matrix columns by sample or group
name.
}
\details{
This will generally not be called directly, but by other modules such as the 
selectmatrix module.
}
\examples{
sampleselectInput(ns('heatmap'))
}
\keyword{shiny}

