% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplot}
\alias{scatterplot}
\title{Server function for the scatterplot module}
\usage{
scatterplot(input, output, session, getDatamatrix, getThreedee = NULL,
  getXAxis = NULL, getYAxis = NULL, getZAxis = NULL,
  getShowLabels = NULL, getPointSize = NULL, getTitle = reactive({     ""
  }), getLabels = reactive({     rownames(getDatamatrix()) }),
  colorby = NULL, size = NULL, allow_3d = TRUE, x = NA, y = NA,
  z = NA, getLines = NULL)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{getDatamatrix}{Reactive supplying a matrix. If using external controls
this should match the one supplied to \code{scatterplotcontrols}}

\item{getThreedee}{A reactive defining whether to plot in 3D. If set to NULL
(default), the \code{scatterplotcontrols} module will be called to create
a set of inputs to supply this value and the axes etc.}

\item{getXAxis}{NULL, or if \code{getThreedee} is a reactive, a reactive 
supplying an integer specifying which column of the matrix supplied by 
\code{getDatamatrix} should be used for this axis.}

\item{getYAxis}{NULL, or if \code{getThreedee} is a reactive, a reactive 
supplying an integer specifying which column of the matrix supplied by 
\code{getDatamatrix} should be used for this axis.}

\item{getZAxis}{NULL, or if \code{getThreedee} is a reactive, a reactive 
supplying an integer specifying which column of the matrix supplied by 
\code{getDatamatrix} should be used for this axis.}

\item{getShowLabels}{NULL, or if \code{getThreedee} is a reactive, a 
reactive supplying a logical defining whether labels should be shown on 
points.}

\item{getPointSize}{NULL, or if \code{getThreedee} is a reactive, a 
reactive supplying an integer point size to pass to plotly.}

\item{colorby}{A reactive returning a factor definining the groups in which 
points should be colored.}

\item{allow_3d}{For case where controls are generated, passed to
\code{scatterplotcontrols}.}

\item{x}{For case where controls are generated, passed to
\code{scatterplotcontrols}.}

\item{y}{For case where controls are generated, passed to
\code{scatterplotcontrols}.}

\item{z}{For case where controls are generated, passed to
\code{scatterplotcontrols}.}

\item{getLines}{Reactive returning a data frame defining lines to be drawn.
Three columns required: name, x and y, with two rows for every value of
name. These two rows represent the start and end of a line.}

\item{title}{A plot title}

\item{labels}{An optional list of labels to use instead of row names from
\code{getDatamatrix()}.}
}
\description{
Controls for this module are provided by the \code{scatterplotcontrols}
module, which is automatically called if reactives are not supplied to the 
server function. This setup allows the same set of controls to power 
multiple scatter plots.
}
\details{
Plots are powered by plotly
}
\examples{
callModule(scatterplot, 'pca', getDatamatrix = pcaMatrix, getThreedee = getThreedee, getXAxis = getXAxis, getYAxis = getYAxis, getZAxis = getZAxis, getShowLabels = getShowLabels, getPointSize = getPointSize, title = 'PCA plot', colorby = na.replace(selectColData()[[colorBy()]], 'N/A'))
}

