% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplotcontrols.R
\name{scatterplotcontrols}
\alias{scatterplotcontrols}
\title{Server function for scatterplotcontrols module}
\usage{
scatterplotcontrols(input, output, session, getDatamatrix, x = NA, y = NA,
  z = NA)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{getDatamatrix}{Reactive expression that returns a matrix from which 
coulumn headers will be used to create axis select drop-downs. The same 
reactive should be supplied to the scatterplot module}

\item{x}{A value supplied for this parameter will cause a hidden field to 
be generated instead of a select, useful for scatter plots that don't need
the user to select axes (default: NA)}

\item{y}{A value supplied for this parameter will cause a hidden field to 
be generated instead of a select, useful for scatter plots that don't need
the user to select axes (default: NA)}

\item{z}{A value supplied for this parameter will cause a hidden field to 
be generated instead of a select, useful for scatter plots that don't need
the user to select axes (default: NA)}
}
\value{
output A list of reactives for accessing input values
}
\description{
This module provides controls (2D/3D, axes etc) for scatter plots, which
may then be used by one or more instances of the scatterplot module.
}
\examples{
unpack.list(callModule(scatterplotcontrols, "pca", pcaMatrix, x = 1, y = 2)) # To have fixed axes rather than user-selected
}

