% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplotcontrols.R
\name{scatterplotcontrolsInput}
\alias{scatterplotcontrolsInput}
\title{Input function for scatterplotcontrols module}
\usage{
scatterplotcontrolsInput(id, allow_3d = TRUE, make_colors = FALSE)
}
\arguments{
\item{id}{Submodule namespace}

\item{allow_3d}{Boolean: allow user to choose 3D plotting?}

\item{make_colors}{Boolean: add controls for coloring?}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
This module provides controls (2D/3D, axes etc) for scatter plots, which
may then be used by one or more instances of the scatterplot module.
}
\examples{
scatterplotcontrolsInput('pca', allow_3d = FALSE) # for a 2D plot
}
