% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectmatrix.R
\name{selectmatrix}
\alias{selectmatrix}
\title{The server function of the selectmatrix module}
\usage{
selectmatrix(input, output, session, se, transcriptfield, entrezgenefield,
  genefield, geneset_files, var_n = 50)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{se}{StructuredExperiment object with assay and experimental data}

\item{transcriptfield}{The main identifier for the rows in the assay data.
This could be transcript ID, but also probe etc.}

\item{entrezgenefield}{The column of annotation containing Entrez gene IDs}

\item{genefield}{The gene ID type in annotation by which results are keyed}

\item{geneset_files}{A named list of .gmt gene set files as might be 
derived from MSigDB}

\item{var_n}{The number of rows to select when doing so by variance. Default = 50}
}
\value{
output A list of reactive functions for fetching the derived matrix 
and making a title based on its properties.
}
\description{
This module uses the geneselect and sampleselect modules to parse controls
defining matrix, row and column and return a subsetted matrix.
}
\details{
This function is not called directly, but rather via callModule() (see 
example).
}
\examples{
selectSamples <- callModule(sampleselect, 'selectmatrix', se)
}
\keyword{shiny}

