% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectmatrix.R
\name{selectmatrix}
\alias{selectmatrix}
\title{The server function of the selectmatrix module}
\usage{
selectmatrix(
  input,
  output,
  session,
  eselist,
  var_n = 50,
  var_max = NULL,
  select_assays = TRUE,
  select_samples = TRUE,
  select_genes = TRUE,
  provide_all_genes = FALSE,
  default_gene_select = NULL,
  require_contrast_stats = FALSE,
  rounding = 2,
  select_meta = TRUE
)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{eselist}{ExploratorySummarizedExperimentList object containing
ExploratorySummarizedExperiment objects}

\item{var_n}{The number of rows to select when doing so by variance. Default
= 50}

\item{var_max}{The maximum umber of rows to select when doing so by variance.
Default = 500}

\item{select_assays}{Provide UI and functions for assay selection?}

\item{select_samples}{Provide UI and functions for sample selection?
(Default: TRUE)}

\item{select_genes}{Provide UI and functions for gene (row) selection? 
(Default: TRUE)}

\item{provide_all_genes}{Allow the 'all rows' selection in the UI? Means we 
don't have to calculate variance so the display is quicker, but it's a bad 
idea for e.g. heatmaps where the visual scales by the numbre of rows.}

\item{default_gene_select}{The default method to use for selecting genes}

\item{require_contrast_stats}{Only use elements of \code{eselist} that have a 
populated \code{contrast_stats} slot. For plots using p value data etc, this is 
used to hide experiments that don't have the necessary data.}

\item{rounding}{Number of decimal places to show in results (Default 2)}

\item{select_meta}{Boolean- add metadata controls?}
}
\value{
output A list of reactive functions for fetching the derived matrix 
  and making a title based on its properties.
}
\description{
This module forms the core of many operations in \code{shinyngs}. To use the
matrix data stored in an \code{ExploratorySummarizedExperimentList} object, 
selection must be made on the basis of experiment (which 
\code{ExploratorySummarizedExperiment}) to use), the specific assay of the 
experiment, and the rows and columns of the selected assay matrix. This 
module provides customisable controls for selection at each of these levels, 
and parses those inputs to produce matrices used in the various plots.
}
\details{
The \code{\link{geneselect}} and \code{\link{sampleselect}} modules provide
row- and column- selection, respectively.
This will generally not be called directly, but by other modules such as the 
heatmap module.

This function is not called directly, but rather via callModule() (see 
example).
}
\examples{
selectSamples <- callModule(sampleselect, 'selectmatrix', eselist)
}
\keyword{shiny}
