% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectmatrix.R
\name{selectmatrix}
\alias{selectmatrix}
\title{The server function of the selectmatrix module}
\usage{
selectmatrix(input, output, session, eses, var_n = 50, var_max = NULL,
  select_samples = TRUE, select_genes = TRUE, provide_all_genes = FALSE,
  require_tests = FALSE, rounding = 2)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{eses}{List of ExploratorySummarizedExperiment objects with assay and 
experimental data}

\item{var_n}{The number of rows to select when doing so by variance. Default
= 50}

\item{var_max}{The maximum umber of rows to select when doing so by variance.
Default = 500}

\item{select_samples}{Provide UI and functions for sample selection?
(Default: TRUE)}

\item{select_genes}{Provide UI and functions for gene (row) selection? 
(Default: TRUE)}

\item{provide_all_genes}{Allow the 'all rows' selection in the UI? Means we 
don't have to calculate variance so the display is quicker, but it's a bad 
idea for e.g. heatmaps where the visual scales by the numbre of rows.}

\item{rounding}{Number of decimal places to show in results (Default 2)}
}
\value{
output A list of reactive functions for fetching the derived matrix 
  and making a title based on its properties.
}
\description{
This module uses the geneselect and sampleselect modules to parse controls 
defining matrix, row and column and return a subsetted matrix.
}
\details{
This function is not called directly, but rather via callModule() (see 
example).
}
\examples{
selectSamples <- callModule(sampleselect, 'selectmatrix', eses)
}
\keyword{shiny}

