% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectmatrix.R
\name{selectmatrixInput}
\alias{selectmatrixInput}
\title{The UI input function of the selectmarix module
 
This module uses the geneselect and sampleselect modules to parse controls
defining matrix, row and column and return a matrix.}
\usage{
selectmatrixInput(id, se, group_vars, default_groupvar)
}
\arguments{
\item{id}{Submodule namespace}

\item{group_vars}{The variables from the structured experiment that should
be used to control sample grouping in the plot}

\item{The}{default grouping variable to use}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
This will generally not be called directly, but by other modules such as the
heatmap module.
}
\examples{
selectmatrixInput(ns("heatmap"), se, group_vars, default_groupvar)
}
\keyword{shiny}

