% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{simpleApp}
\alias{simpleApp}
\title{Produce a simple app with controls and layout for a single module, in a
shiny \code{sideBarLayout()}.}
\usage{
simpleApp(eses, module = NULL, title = NULL, ui_only = FALSE, ...)
}
\arguments{
\item{module}{Character string specifying the module to use}

\item{title}{An optional title string to use with the module name}

\item{ses}{List of ExploratorySummarizedExperiment objects with assay and experimental
data}
}
\description{
Internal function called by prepareApp() to make simple single-function 
apps.
}
\examples{
simpleApp(ses, 'heatmap', 'My study name')
}
\keyword{shiny}

