% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpletable.R
\name{simpletable}
\alias{simpletable}
\title{The server function of the simpletable module}
\usage{
simpletable(
  id,
  downloadMatrix = NULL,
  displayMatrix,
  pageLength = 15,
  filename,
  rownames = FALSE,
  show_controls = TRUE,
  filter = "none"
)
}
\arguments{
\item{id}{Module namespace}

\item{downloadMatrix}{Reactive expression for retrieving the plot to supply
for download (default: NULL, in which case \code{displayMatrix()} will be
used)}

\item{displayMatrix}{Reactive expression for retrieving the plot for display}

\item{pageLength}{Number of items per page}

\item{filename}{A string to use in the name of the table download}

\item{rownames}{Passed to \code{\link[DT]{renderDataTable}}. Display the row
names of the input matrix? (default: FALSE)}

\item{show_controls}{Show search box, controls etc on resulting data table?
(default: TRUE)}

\item{filter}{Passed to \code{DT::renderDataTable()}}
}
\description{
This function is called directly, using the same id as its UI counterpart,
and wraps its logic in \code{moduleServer()} (see example).
}
\examples{
simpletable("simpletable", my_data_frame)

}
\keyword{shiny}
