% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpletable.R
\name{simpletable}
\alias{simpletable}
\title{The server function of the simpletable module}
\usage{
simpletable(input, output, session, datatable, pageLength = 25)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{datatable}{Data frame with a table of data}

\item{pageLength}{Number of items per page}
}
\description{
This function is not called directly, but rather via callModule() (see 
example).
}
\details{
This function assumes that the gene sets have one gene ID (e.g. Entrez)
which need to be converted to another (e.g. Symbol, Ensembl gene ID).
This would be common when dealign with MSigDB gene sets, for example.
}
\examples{
callModule(simpletable, 'simpletable', my_data_frame)
}
\keyword{shiny}

