% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpletable.R
\name{simpletable}
\alias{simpletable}
\title{The server function of the simpletable module}
\usage{
simpletable(input, output, session, getDatatable, pageLength = 25,
  filename = "datatable")
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{pageLength}{Number of items per page}

\item{filename}{A string to use in the name of the table download}

\item{datatable}{Data frame with a table of data}
}
\description{
This function is not called directly, but rather via callModule() (see 
example).
}
\examples{
callModule(simpletable, 'simpletable', my_data_frame)
}
\keyword{shiny}

