% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpletable.R
\name{simpletableInput}
\alias{simpletableInput}
\title{The UI input function of the simpletable module
 
This module produces a simple datatable output with a download button.}
\usage{
simpletableInput(id, tabletitle = "", description = NULL)
}
\arguments{
\item{id}{Submodule namespace}

\item{tabletitle}{Table title. Will be used on download button.}

\item{description}{A string to display in the side bar to explain the table.}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
The UI input function of the simpletable module
 
This module produces a simple datatable output with a download button.
}
\examples{
simpletableInput('mytable', 'this is a table')
}
\keyword{shiny}
