% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpletable.R
\name{simpletableOutput}
\alias{simpletableOutput}
\title{The output function of the simpletable module}
\usage{
simpletableOutput(id, tabletitle = NULL)
}
\arguments{
\item{id}{Module namespace}

\item{title}{(optional) Title to display with the table}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
This provides actual datatable element for display by applications
}
\examples{
simpletableOutput('simpletable', 'my title')
}
\keyword{shiny}
