% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarisematrix.R
\name{summarisematrix}
\alias{summarisematrix}
\title{The server function of the summarisematrix module}
\usage{
summarisematrix(input, output, session)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}
}
\description{
This module provides a form element and associated get function for defining
how a summary statistic is calculated (probably by mean).
}
\examples{
callModule(summarisematrix)
}
\keyword{shiny}

