% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Call the various read/ validate methods for input data surrounding an experiment}
\usage{
validate_inputs(
  samples_metadata,
  features_metadata,
  assay_files,
  contrasts_file,
  sample_id_col,
  feature_id_col,
  assay_names = NULL,
  differential_results = NULL,
  feature_id_column = "gene_id",
  pval_column = "pval_column",
  qval_column = "qval_column",
  fc_column = "log2FoldChange",
  unlog_foldchanges = FALSE
)
}
\arguments{
\item{samples_metadata}{Sample metadata data frame}

\item{features_metadata}{Feature metadata data frame}

\item{assay_files}{List of assay matrices}

\item{contrasts_file}{Contrasts definition file}

\item{sample_id_col}{Column of sample metadata used for identifiers}

\item{feature_id_col}{Column of feature metadata used for identifiers}
}
\value{
output A named list with feature/ observation components
}
\description{
Call the various read/ validate methods for input data surrounding an experiment
}
