% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoplot.R
\name{volcanoplotInput}
\alias{volcanoplotInput}
\title{The UI input function of the volcanoplot module
 
This module provides information on the comparison betwen pairs of groups 
defined in a 'tests' slot of the SummarizedExperiment metadata.}
\usage{
volcanoplotInput(id, ses)
}
\arguments{
\item{id}{Submodule namespace}

\item{ses}{List of structuredExperiment objects with assay and experimental
data, with additional information in the metadata() slot}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
Leverages the \code{scatterplot} module
}
\examples{
differentialtableInput('experiment', ses)
}
\keyword{shiny}

