% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoplot.R
\name{volcanoplotInput}
\alias{volcanoplotInput}
\title{The UI input function of the volcanoplot module}
\usage{
volcanoplotInput(id, eses)
}
\arguments{
\item{id}{Submodule namespace}

\item{ses}{List of ExploratorySummarizedExperiment objects with assay and 
experimental data}
}
\value{
output An HTML tag object that can be rendered as HTML using 
  as.character()
}
\description{
This module provides information on the comparison betwen pairs of groups 
defined in a 'tests' slot of the ExploratorySummarizedExperiment
}
\details{
Leverages the \code{scatterplot} module
}
\examples{
differentialtableInput('experiment', eses)
}
\keyword{shiny}

