% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoplot.R
\name{volcanoplotOutput}
\alias{volcanoplotOutput}
\title{The output function of the \code{volcanoplot} module}
\usage{
volcanoplotOutput(id)
}
\arguments{
\item{id}{Module namespace}
}
\value{
output An HTML tag object that can be rendered as HTML using 
as.character()
}
\description{
A volcano plot displays -log(10) of a p value/ FDR against a log(2) fold 
change on the x axis. This module produces such a plot using the 
\code{\link{scatterplot}} module (which uses 
\code{\link[plotly]{plot_ly}})), using data provided by the 
\code{\link{contrasts}} module based on the setting of \code{contrasts} in 
\code{eselist}.
}
\details{
Leverages the \code{scatterplot} module
}
\examples{
differentialtableOutput('experiment')

# However, almost certainly called via application creation:

data(zhangneurons)
app <- prepareApp('volcanoplot', zhangneurons)
shinyApp(ui = app$ui, server = app$server)
}
\keyword{shiny}

